(test(dev-tools)): split oversized dev-tools tests and record QA evidence
- Split large Pytest modules into focused parts to keep each file <=500 lines while preserving collection
- Adjust test guard fixture scoping to match new module prefixes after the split
- Capture remediation QA gate artifacts (Python/PowerShell/TypeScript + line-count verification) and mark plan complete

Refs: #40
