+++ docs/features/active/2026-02-21-bootstrap-utility-scripts-40/evidence/other/remediation-closeout.2026-02-21T22-24.md
@@ -1,0 +1,13 @@
+# Remediation Closeout — bootstrap-utility-scripts (#40)
+
+Timestamp: 2026-02-22T02:08:00-05:00
+AC1Status: MET
+AC4Status: MET
+AC5Status: MET
+PreflightSignal: PREFLIGHT: ALL CLEAR
+
+Summary:
+- Python type-check scope remained green with `poetry run pyright` exiting 0.
+- Oversized test files were split into focused modules, all ≤500 lines, with pytest collection and execution preserved.
+- Required QA evidence artifacts were recorded for Python, PowerShell, TypeScript, and line-count verification under `evidence/qa-gates/`.
+- Final full QA loops completed successfully for Python, PowerShell, and TypeScript.
