+++ docs/features/active/2026-02-21-bootstrap-utility-scripts-40/evidence/qa-gates/line-counts.2026-02-21T22-24.md
@@ -1,0 +1,33 @@
+# Line Count Verification Evidence
+
+Timestamp: 2026-02-22T01:22:00-05:00
+Command: poetry run python - (status-porcelain tracked `.py/.ps1/.psm1/.psd1/.ts` line-count verifier)
+EXIT_CODE: 0
+Output Summary:
+- 20 changed tracked code/test/script files verified
+- Highest line count in modified files: 479
+- Violations (>500 lines): 0
+
+Verified file counts:
+- tests/scripts/dev_tools/atomic_executor/test_cli.py: 405
+- tests/scripts/dev_tools/atomic_executor/test_cli_part2.py: 458
+- tests/scripts/dev_tools/atomic_executor/test_cli_part2_part2.py: 143
+- tests/scripts/dev_tools/atomic_executor/test_cli_part3.py: 478
+- tests/scripts/dev_tools/atomic_executor/test_cli_part4.py: 424
+- tests/scripts/dev_tools/atomic_executor/test_cli_part4_part2.py: 285
+- tests/scripts/dev_tools/test_collect_pr_context.py: 462
+- tests/scripts/dev_tools/test_collect_pr_context_part2.py: 463
+- tests/scripts/dev_tools/test_collect_pr_context_part3.py: 460
+- tests/scripts/dev_tools/test_collect_pr_context_part4.py: 337
+- tests/scripts/dev_tools/test_github.py: 454
+- tests/scripts/dev_tools/test_github_part2.py: 372
+- tests/scripts/dev_tools/test_github_part3.py: 186
+- tests/scripts/dev_tools/test_new_active_feature_folder.py: 462
+- tests/scripts/dev_tools/test_new_active_feature_folder_part2.py: 469
+- tests/scripts/dev_tools/test_new_active_feature_folder_part3.py: 459
+- tests/scripts/dev_tools/test_new_active_feature_folder_part4.py: 366
+- tests/scripts/dev_tools/test_resolve_execute_plan_prompt.py: 479
+- tests/scripts/dev_tools/test_resolve_execute_plan_prompt_part2.py: 459
+- tests/scripts/dev_tools/test_resolve_execute_plan_prompt_part3.py: 240
+
+GateStatus: PASS
