+++ docs/features/active/2026-02-21-bootstrap-utility-scripts-40/evidence/qa-gates/typescript.2026-02-21T22-24.md
@@ -1,0 +1,57 @@
+# TypeScript QA Gate Evidence
+
+## Run 1 — Targeted verification
+
+Timestamp: 2026-02-22T01:33:00-05:00
+Command: npm run format:check
+EXIT_CODE: 0
+Output Summary:
+- Prettier check passed; all matched files are formatted
+
+Timestamp: 2026-02-22T01:33:20-05:00
+Command: npm run lint
+EXIT_CODE: 0
+Output Summary:
+- ESLint completed without errors
+
+Timestamp: 2026-02-22T01:33:40-05:00
+Command: npm run typecheck
+EXIT_CODE: 0
+Output Summary:
+- TypeScript compiler `--noEmit` completed without diagnostics
+
+Timestamp: 2026-02-22T01:34:00-05:00
+Command: npm run test:unit
+EXIT_CODE: 0
+Output Summary:
+- Jest: 1 passed, 0 failed
+
+GateStatus: PASS
+
+## Run 2 — Final full toolchain loop (single uninterrupted clean pass)
+
+Timestamp: 2026-02-22T02:02:00-05:00
+Command: npm run format
+EXIT_CODE: 0
+Output Summary:
+- Prettier write pass completed; all matched files unchanged
+
+Timestamp: 2026-02-22T02:02:20-05:00
+Command: npm run lint
+EXIT_CODE: 0
+Output Summary:
+- ESLint completed without errors
+
+Timestamp: 2026-02-22T02:02:40-05:00
+Command: npm run typecheck
+EXIT_CODE: 0
+Output Summary:
+- TypeScript compiler `--noEmit` completed without diagnostics
+
+Timestamp: 2026-02-22T02:03:00-05:00
+Command: npm run test:unit
+EXIT_CODE: 0
+Output Summary:
+- Jest: 1 passed, 0 failed
+
+GateStatus: PASS
