+++ tests/scripts/dev_tools/atomic_executor/test_cli.py
@@ -7,7 +7,6 @@ clipboard operations, and main execution orchestration.
 
 # pyright: reportArgumentType=false, reportUnknownLambdaType=false, reportUnknownArgumentType=false, reportPrivateUsage=false
 
-import os
 import subprocess
 import sys
 from pathlib import Path
@@ -400,1602 +399,3 @@ class TestCopyToClipboard:
 
         assert result is True
         assert "/usr/bin/xclip" in attempted_commands
-
-
-class TestMainEdgeCases:
-    """Edge case tests for main execution flow."""
-
-    def test_main_exits_early_with_print_prompt(
-        self,
-        mem_path: Path,
-        monkeypatch: "MonkeyPatch",
-        capsys: pytest.CaptureFixture[str],
-    ) -> None:
-        """main() exits early with --print-prompt without running copilot."""
-        from scripts.dev_tools.atomic_executor.cli import main
-
-        # Setup minimal feature folder
-        feature_dir = mem_path / "docs" / "features" / "active" / "my-feature"
-        feature_dir.mkdir(parents=True)
-        (feature_dir / "plan.md").write_text(
-            "# Phase 0\n- [ ] [P0-T1] Task 1\n\n"
-            "# Phase 2 (QA/Toolchain)\n"
-            "- [ ] [P2-T1] Black\n"
-            "- [ ] [P2-T2] Ruff\n"
-            "- [ ] [P2-T3] Pyright\n"
-            "- [ ] [P2-T4] Pytest",
-            encoding="utf-8",
-        )
-        (feature_dir / "spec.md").write_text("# Spec\n", encoding="utf-8")
-
-        template_dir = mem_path / ".github" / "prompts"
-        template_dir.mkdir(parents=True)
-        (template_dir / "execute-plan-template.md").write_text(
-            "TEMPLATE\n", encoding="utf-8"
-        )
-
-        # Mock all subprocess calls
-        def mock_run(*args: object, **kwargs: object) -> Mock:
-            result = Mock()
-            result.stdout = ""
-            result.returncode = 0
-            return result  # type: ignore[return-value]
-
-        monkeypatch.setattr("subprocess.run", mock_run)
-
-        # Mock copilot to not be found (shouldn't be called)
-        monkeypatch.setattr("shutil.which", lambda x: None if x == "copilot" else "git")  # type: ignore[arg-type,misc]
-
-        exit_code = main(
-            [
-                "execute",
-                str(feature_dir),
-                "--workspace",
-                str(mem_path),
-                "--print-prompt",
-                "--skip-preflight-qc",
-            ]
-        )
-
-        assert exit_code == 0
-        captured = capsys.readouterr()
-        assert "TEMPLATE" in captured.out
-        assert "CURRENT TASK" in captured.out
-
-    def test_main_exits_early_with_copy_prompt(
-        self,
-        mem_path: Path,
-        monkeypatch: "MonkeyPatch",
-        capsys: pytest.CaptureFixture[str],
-    ) -> None:
-        """main() exits early with --copy-prompt without running copilot."""
-        from scripts.dev_tools.atomic_executor.cli import main
-
-        # Setup minimal feature folder
-        feature_dir = mem_path / "docs" / "features" / "active" / "my-feature"
-        feature_dir.mkdir(parents=True)
-        (feature_dir / "plan.md").write_text(
-            "# Phase 0\n- [ ] [P0-T1] Task 1\n\n"
-            "# Phase 2 (QA/Toolchain)\n"
-            "- [ ] [P2-T1] Black\n"
-            "- [ ] [P2-T2] Ruff\n"
-            "- [ ] [P2-T3] Pyright\n"
-            "- [ ] [P2-T4] Pytest",
-            encoding="utf-8",
-        )
-        (feature_dir / "spec.md").write_text("# Spec\n", encoding="utf-8")
-
-        template_dir = mem_path / ".github" / "prompts"
-        template_dir.mkdir(parents=True)
-        (template_dir / "execute-plan-template.md").write_text(
-            "TEMPLATE\n", encoding="utf-8"
-        )
-
-        # Mock all subprocess calls
-        def mock_run(*args: object, **kwargs: object) -> Mock:
-            result = Mock()
-            result.stdout = ""
-            result.returncode = 0
-            return result  # type: ignore[return-value]
-
-        monkeypatch.setattr("subprocess.run", mock_run)
-
-        # Mock clipboard to succeed
-        monkeypatch.setattr(
-            "scripts.dev_tools.atomic_executor.cli.copy_to_clipboard", lambda x: True  # type: ignore[arg-type,misc]
-        )
-
-        exit_code = main(
-            [
-                "execute",
-                str(feature_dir),
-                "--workspace",
-                str(mem_path),
-                "--copy-prompt",
-                "--skip-preflight-qc",
-            ]
-        )
-
-        assert exit_code == 0
-        captured = capsys.readouterr()
-        assert "copied to clipboard" in captured.err.lower()
-
-    def test_main_returns_error_for_missing_plan(
-        self,
-        mem_path: Path,
-        monkeypatch: "MonkeyPatch",
-        capsys: pytest.CaptureFixture[str],
-    ) -> None:
-        """main() returns error code when plan.md missing."""
-        from scripts.dev_tools.atomic_executor.cli import main
-
-        # Setup feature folder without plan.md
-        feature_dir = mem_path / "docs" / "features" / "active" / "my-feature"
-        feature_dir.mkdir(parents=True)
-        (feature_dir / "spec.md").write_text("# Spec\n", encoding="utf-8")
-
-        # Mock git to be clean
-        def mock_run(*args: object, **kwargs: object) -> Mock:
-            result = Mock()
-            result.stdout = ""
-            result.returncode = 0
-            return result  # type: ignore[return-value]
-
-        monkeypatch.setattr("subprocess.run", mock_run)
-
-        exit_code = main(
-            [
-                "execute",
-                str(feature_dir),
-                "--workspace",
-                str(mem_path),
-            ]
-        )
-
-        assert exit_code == 2
-        captured = capsys.readouterr()
-        assert "Missing required plan file" in captured.err
-
-    def test_main_returns_zero_when_plan_already_complete(
-        self,
-        mem_path: Path,
-        monkeypatch: "MonkeyPatch",
-        capsys: pytest.CaptureFixture[str],
-    ) -> None:
-        """main() returns 0 when all tasks already checked."""
-        from scripts.dev_tools.atomic_executor.cli import main
-
-        # Setup feature folder with all tasks checked
-        feature_dir = mem_path / "docs" / "features" / "active" / "my-feature"
-        feature_dir.mkdir(parents=True)
-        (feature_dir / "plan.md").write_text(
-            "# Phase 0\n- [x] [P0-T1] Task 1\n\n"
-            "# Phase 2 (QA/Toolchain)\n"
-            "- [x] [P2-T1] Black\n"
-            "- [x] [P2-T2] Ruff\n"
-            "- [x] [P2-T3] Pyright\n"
-            "- [x] [P2-T4] Pytest",
-            encoding="utf-8",
-        )
-        (feature_dir / "spec.md").write_text("# Spec\n", encoding="utf-8")
-
-        template_dir = mem_path / ".github" / "prompts"
-        template_dir.mkdir(parents=True)
-        (template_dir / "execute-plan-template.md").write_text(
-            "TEMPLATE\n", encoding="utf-8"
-        )
-
-        # Mock git to be clean
-        def mock_run(*args: object, **kwargs: object) -> Mock:
-            result = Mock()
-            result.stdout = ""
-            result.returncode = 0
-            return result  # type: ignore[return-value]
-
-        monkeypatch.setattr("subprocess.run", mock_run)
-
-        exit_code = main(
-            [
-                "resume",
-                str(feature_dir),
-                "--workspace",
-                str(mem_path),
-            ]
-        )
-
-        assert exit_code == 0
-        captured = capsys.readouterr()
-        assert "already complete" in captured.out.lower()
-
-    def test_main_returns_error_for_missing_template(
-        self,
-        mem_path: Path,
-        monkeypatch: "MonkeyPatch",
-        capsys: pytest.CaptureFixture[str],
-    ) -> None:
-        """main() returns error code when prompt template missing."""
-        from scripts.dev_tools.atomic_executor.cli import main
-
-        # Setup feature folder with plan.md
-        feature_dir = mem_path / "docs" / "features" / "active" / "my-feature"
-        feature_dir.mkdir(parents=True)
-        (feature_dir / "plan.md").write_text(
-            "# Phase 0\n- [ ] [P0-T1] Task 1\n\n"
-            "# Phase 2 (QA/Toolchain)\n"
-            "- [ ] [P2-T1] Black\n"
-            "- [ ] [P2-T2] Ruff\n"
-            "- [ ] [P2-T3] Pyright\n"
-            "- [ ] [P2-T4] Pytest",
-            encoding="utf-8",
-        )
-        (feature_dir / "spec.md").write_text("# Spec\n", encoding="utf-8")
-
-        # Don't create template file
-
-        # Mock git to be clean
-        def mock_run(*args: object, **kwargs: object) -> Mock:
-            result = Mock()
-            result.stdout = ""
-            result.returncode = 0
-            return result  # type: ignore[return-value]
-
-        monkeypatch.setattr("subprocess.run", mock_run)
-
-        exit_code = main(
-            [
-                "execute",
-                str(feature_dir),
-                "--workspace",
-                str(mem_path),
-            ]
-        )
-
-        assert exit_code == 2
-        captured = capsys.readouterr()
-        assert "Prompt template not found" in captured.err
-
-    def test_main_with_copy_prompt_fallback_when_clipboard_fails(
-        self,
-        mem_path: Path,
-        monkeypatch: "MonkeyPatch",
-        capsys: pytest.CaptureFixture[str],
-    ) -> None:
-        """main() prints prompt when --copy-prompt fails."""
-        from scripts.dev_tools.atomic_executor.cli import main
-
-        # Setup minimal feature folder
-        feature_dir = mem_path / "docs" / "features" / "active" / "my-feature"
-        feature_dir.mkdir(parents=True)
-        (feature_dir / "plan.md").write_text(
-            "# Phase 0\n- [ ] [P0-T1] Task 1\n\n"
-            "# Phase 2 (QA/Toolchain)\n"
-            "- [ ] [P2-T1] Black\n"
-            "- [ ] [P2-T2] Ruff\n"
-            "- [ ] [P2-T3] Pyright\n"
-            "- [ ] [P2-T4] Pytest",
-            encoding="utf-8",
-        )
-        (feature_dir / "spec.md").write_text("# Spec\n", encoding="utf-8")
-
-        template_dir = mem_path / ".github" / "prompts"
-        template_dir.mkdir(parents=True)
-        (template_dir / "execute-plan-template.md").write_text(
-            "TEMPLATE\n", encoding="utf-8"
-        )
-
-        # Mock all subprocess calls
-        def mock_run(*args: object, **kwargs: object) -> Mock:
-            result = Mock()
-            result.stdout = ""
-            result.returncode = 0
-            return result  # type: ignore[return-value]
-
-        monkeypatch.setattr("subprocess.run", mock_run)
-
-        # Mock clipboard to fail
-        monkeypatch.setattr(
-            "scripts.dev_tools.atomic_executor.cli.copy_to_clipboard",
-            lambda x: False,  # type: ignore[arg-type,misc]
-        )
-
-        exit_code = main(
-            [
-                "execute",
-                str(feature_dir),
-                "--workspace",
-                str(mem_path),
-                "--copy-prompt",
-                "--skip-preflight-qc",
-            ]
-        )
-
-        assert exit_code == 0
-        captured = capsys.readouterr()
-        assert "Clipboard copy not available" in captured.err
-        assert "TEMPLATE" in captured.out
-
-    def test_main_execute_with_start_flag(
-        self,
-        mem_path: Path,
-        monkeypatch: "MonkeyPatch",
-        capsys: pytest.CaptureFixture[str],
-    ) -> None:
-        """main() executes with --start flag to begin at specific task."""
-        from scripts.dev_tools.atomic_executor.cli import main
-
-        # Setup feature folder with multiple tasks
-        feature_dir = mem_path / "docs" / "features" / "active" / "my-feature"
-        feature_dir.mkdir(parents=True)
-        (feature_dir / "plan.md").write_text(
-            "# Phase 0\n"
-            "- [x] [P0-T1] First task\n"
-            "- [ ] [P0-T2] Second task\n\n"
-            "# Phase 2 (QA/Toolchain)\n"
-            "- [ ] [P2-T1] Black\n"
-            "- [ ] [P2-T2] Ruff\n"
-            "- [ ] [P2-T3] Pyright\n"
-            "- [ ] [P2-T4] Pytest",
-            encoding="utf-8",
-        )
-        (feature_dir / "spec.md").write_text("# Spec\n", encoding="utf-8")
-
-        template_dir = mem_path / ".github" / "prompts"
-        template_dir.mkdir(parents=True)
-        (template_dir / "execute-plan-template.md").write_text(
-            "TEMPLATE\n", encoding="utf-8"
-        )
-
-        # Mock git to be clean
-        def mock_run(*args: object, **kwargs: object) -> Mock:
-            result = Mock()
-            result.stdout = ""
-            result.returncode = 0
-            return result  # type: ignore[return-value]
-
-        monkeypatch.setattr("subprocess.run", mock_run)
-
-        exit_code = main(
-            [
-                "execute",
-                str(feature_dir),
-                "--workspace",
-                str(mem_path),
-                "--start",
-                "P0-T2",
-                "--print-prompt",
-                "--skip-preflight-qc",
-            ]
-        )
-
-        assert exit_code == 0
-        captured = capsys.readouterr()
-        assert "P0-T2" in captured.out
-        assert "Second task" in captured.out
-
-    def test_main_execute_when_all_tasks_complete(
-        self,
-        mem_path: Path,
-        monkeypatch: "MonkeyPatch",
-        capsys: pytest.CaptureFixture[str],
-    ) -> None:
-        """main() exits when execute subcommand finds no unchecked tasks."""
-        from scripts.dev_tools.atomic_executor.cli import main
-
-        # Setup feature folder with all tasks checked
-        feature_dir = mem_path / "docs" / "features" / "active" / "my-feature"
-        feature_dir.mkdir(parents=True)
-        (feature_dir / "plan.md").write_text(
-            "# Phase 0\n- [x] [P0-T1] Task 1\n\n"
-            "# Phase 2 (QA/Toolchain)\n"
-            "- [x] [P2-T1] Black\n"
-            "- [x] [P2-T2] Ruff\n"
-            "- [x] [P2-T3] Pyright\n"
-            "- [x] [P2-T4] Pytest",
-            encoding="utf-8",
-        )
-        (feature_dir / "spec.md").write_text("# Spec\n", encoding="utf-8")
-
-        template_dir = mem_path / ".github" / "prompts"
-        template_dir.mkdir(parents=True)
-        (template_dir / "execute-plan-template.md").write_text(
-            "TEMPLATE\n", encoding="utf-8"
-        )
-
-        # Mock git to be clean
-        def mock_run(*args: object, **kwargs: object) -> Mock:
-            result = Mock()
-            result.stdout = ""
-            result.returncode = 0
-            return result  # type: ignore[return-value]
-
-        monkeypatch.setattr("subprocess.run", mock_run)
-
-        exit_code = main(
-            [
-                "execute",
-                str(feature_dir),
-                "--workspace",
-                str(mem_path),
-            ]
-        )
-
-        assert exit_code == 0
-        captured = capsys.readouterr()
-        assert "already complete" in captured.out.lower()
-
-    def test_main_successful_execution_with_scoped_qc(
-        self,
-        mem_path: Path,
-        monkeypatch: "MonkeyPatch",
-        capsys: pytest.CaptureFixture[str],
-    ) -> None:
-        """main() successfully executes task with scoped QC."""
-        from scripts.dev_tools.atomic_executor.cli import main
-
-        monkeypatch.setenv("XDG_CONFIG_HOME", str(mem_path / "config-root"))
-
-        # Setup feature folder
-        feature_dir = mem_path / "docs" / "features" / "active" / "my-feature"
-        feature_dir.mkdir(parents=True)
-        plan_file = feature_dir / "plan.md"
-        plan_file.write_text(
-            "# Phase 0\n- [ ] [P0-T1] Task 1\n\n"
-            "# Phase 2 (QA/Toolchain)\n"
-            "- [ ] [P2-T1] Black\n"
-            "- [ ] [P2-T2] Ruff\n"
-            "- [ ] [P2-T3] Pyright\n"
-            "- [ ] [P2-T4] Pytest",
-            encoding="utf-8",
-        )
-        (feature_dir / "spec.md").write_text("# Spec\n", encoding="utf-8")
-
-        template_dir = mem_path / ".github" / "prompts"
-        template_dir.mkdir(parents=True)
-        (template_dir / "execute-plan-template.md").write_text(
-            "Task: {{task_id}}\n", encoding="utf-8"
-        )
-
-        # Setup fake copilot on PATH for run_copilot
-        bin_dir = mem_path / "bin"
-        bin_dir.mkdir()
-        copilot_exe = bin_dir / "copilot"
-        copilot_exe.touch()
-        copilot_exe.chmod(0o755)  # Make executable-ish
-        path = os.environ.get("PATH", "")
-        monkeypatch.setenv("PATH", f"{str(bin_dir)}{os.pathsep}{path}")
-
-        # Track subprocess calls
-        subprocess_calls: list[list[str]] = []
-
-        def mock_run(
-            argv: list[str], *args: object, **kwargs: object
-        ) -> subprocess.CompletedProcess[str]:
-            subprocess_calls.append(argv)
-            result = Mock()
-            result.stdout = ""
-            result.returncode = 0
-            return result  # type: ignore[return-value]
-
-        class MockStdout:
-            def read(self, size: int = -1) -> bytes:
-                return b""
-
-        class MockPopen:
-            def __init__(
-                self, argv: list[str], *args: object, **kwargs: object
-            ) -> None:
-                subprocess_calls.append(argv)
-                self.stdout = MockStdout()
-                self.returncode = 0
-
-            def poll(self) -> int:
-                return 0
-
-            def wait(self) -> int:
-                return 0
-
-        monkeypatch.setattr("subprocess.run", mock_run)
-        monkeypatch.setattr("subprocess.Popen", MockPopen)
-        monkeypatch.setattr("shutil.which", lambda x: f"/usr/bin/{x}")
-
-        # Mock QCRunner methods to succeed
-        from scripts.dev_tools.atomic_executor.qc_runner import QCRunner
-
-        monkeypatch.setattr(
-            QCRunner,
-            "run_scoped",
-            lambda self, expectations=None: None,  # Accept expectations param
-        )
-
-        exit_code = main(
-            [
-                "execute",
-                str(feature_dir),
-                "--workspace",
-                str(mem_path),
-                "--skip-preflight-qc",
-            ]
-        )
-
-        assert exit_code == 0
-        captured = capsys.readouterr()
-        assert "complete and gated" in captured.out.lower()
-
-        # Verify copilot was invoked
-        copilot_calls = [c for c in subprocess_calls if "copilot" in c[0]]
-        assert len(copilot_calls) == 1
-
-
-class TestPreflightQC:
-    """Tests for pre-flight QC functionality."""
-
-    def test_preflight_qc_result_dataclass(self) -> None:
-        """PreflightQCResult stores success status and output."""
-        from scripts.dev_tools.atomic_executor.cli import PreflightQCResult
-        from scripts.dev_tools.atomic_executor.qc_toolchain import QCToolchain
-
-        # Success case
-        result = PreflightQCResult(success=True, output="All passed")
-        assert result.success is True
-        assert result.output == "All passed"
-        assert result.failed_step is None
-        assert result.toolchain == QCToolchain.PYTHON
-
-        # Failure case
-        result = PreflightQCResult(
-            success=False, output="Ruff failed", failed_step="ruff"
-        )
-        assert result.success is False
-        assert result.output == "Ruff failed"
-        assert result.failed_step == "ruff"
-        assert result.toolchain == QCToolchain.PYTHON
-
-    def test_build_preflight_qc_fix_prompt_includes_workspace(
-        self,
-        mem_path: Path,
-    ) -> None:
-        """_build_preflight_qc_fix_prompt includes workspace in prompt."""
-        from scripts.dev_tools.atomic_executor.cli import _build_preflight_qc_fix_prompt
-        from scripts.dev_tools.atomic_executor.qc_toolchain import QCToolchain
-
-        prompt = _build_preflight_qc_fix_prompt(
-            workspace=mem_path,
-            qc_output="Black failed: file.py",
-            toolchain=QCToolchain.PYTHON,
-        )
-
-        # Check key elements are present
-        assert "Pre-flight QC Fix Required" in prompt
-        assert str(mem_path.as_posix()) in prompt
-        assert "Black failed: file.py" in prompt
-        assert "poetry run black" in prompt
-        assert "poetry run ruff" in prompt
-        assert "poetry run pyright" in prompt
-        assert "poetry run pytest" in prompt
-        assert "Do NOT end your turn until all QC steps pass" in prompt
-
-    def test_build_preflight_qc_fix_prompt_typescript(self) -> None:
-        """_build_preflight_qc_fix_prompt uses npm commands for TypeScript."""
-        from scripts.dev_tools.atomic_executor.cli import _build_preflight_qc_fix_prompt
-        from scripts.dev_tools.atomic_executor.qc_toolchain import QCToolchain
-
-        prompt = _build_preflight_qc_fix_prompt(
-            workspace=Path.cwd(),
-            qc_output="npm run test:unit failed",
-            toolchain=QCToolchain.TYPESCRIPT,
-        )
-
-        assert "npm run format" in prompt
-        assert "npm run lint" in prompt
-        assert "npm run typecheck" in prompt
-        assert "npm run test:unit" in prompt
-
-    def test_run_preflight_qc_with_capture_returns_success(
-        self,
-        mem_path: Path,
-        monkeypatch: "MonkeyPatch",
-    ) -> None:
-        """_run_preflight_qc_with_capture returns success when all steps pass."""
-        from scripts.dev_tools.atomic_executor.cli import (
-            _run_preflight_qc_with_capture,
-        )
-
-        # Mock subprocess.run to always succeed
-        def mock_run(
-            *args: object, **kwargs: object
-        ) -> subprocess.CompletedProcess[str]:
-            return subprocess.CompletedProcess(
-                args=args,
-                returncode=0,
-                stdout="OK",
-                stderr="",
-            )
-
-        monkeypatch.setattr("subprocess.run", mock_run)
-
-        result = _run_preflight_qc_with_capture(mem_path)
-
-        assert result.success is True
-        assert result.failed_step is None
-        assert "=== BLACK ===" in result.output
-        assert "=== RUFF ===" in result.output
-        assert "=== PYRIGHT ===" in result.output
-        assert "=== PYTEST ===" in result.output
-
-    def test_run_preflight_qc_with_capture_returns_failure(
-        self,
-        mem_path: Path,
-        monkeypatch: "MonkeyPatch",
-    ) -> None:
-        """_run_preflight_qc_with_capture returns failure when a step fails."""
-        from scripts.dev_tools.atomic_executor.cli import (
-            _run_preflight_qc_with_capture,
-        )
-
-        call_count = 0
-
-        # Mock subprocess.run to fail on second step (ruff)
-        def mock_run(
-            *args: object, **kwargs: object
-        ) -> subprocess.CompletedProcess[str]:
-            nonlocal call_count
-            call_count += 1
-            if call_count == 1:  # black
-                return subprocess.CompletedProcess(
-                    args=args, returncode=0, stdout="Black OK", stderr=""
-                )
-            # ruff fails
-            return subprocess.CompletedProcess(
-                args=args, returncode=1, stdout="Ruff error", stderr="E501"
-            )
-
-        monkeypatch.setattr("subprocess.run", mock_run)
-
-        result = _run_preflight_qc_with_capture(mem_path)
-
-        assert result.success is False
-        assert result.failed_step == "ruff"
-        assert "=== BLACK ===" in result.output
-        assert "=== RUFF ===" in result.output
-        assert "Ruff error" in result.output
-        # Should not have pyright/pytest since ruff failed
-        assert "=== PYRIGHT ===" not in result.output
-
-    def test_run_preflight_qc_with_capture_handles_missing_executable(
-        self,
-        monkeypatch: "MonkeyPatch",
-    ) -> None:
-        """_run_preflight_qc_with_capture fails fast when an executable is missing."""
-        from scripts.dev_tools.atomic_executor.cli import (
-            MISSING_EXECUTABLE_PREFIX,
-            _run_preflight_qc_with_capture,
-        )
-        from scripts.dev_tools.atomic_executor.qc_toolchain import QCToolchain
-
-        monkeypatch.setattr("shutil.which", lambda _cmd: None)
-
-        def _should_not_run(*args: object, **kwargs: object) -> None:
-            pytest.fail("subprocess.run should not be called when executable missing")
-
-        monkeypatch.setattr("subprocess.run", _should_not_run)
-
-        result = _run_preflight_qc_with_capture(
-            Path.cwd(), toolchain=QCToolchain.TYPESCRIPT
-        )
-
-        assert result.success is False
-        assert result.failed_step == "format"
-        assert MISSING_EXECUTABLE_PREFIX in result.output
-
-    def test_skip_preflight_qc_flag_parses(self) -> None:
-        """parse_args() parses --skip-preflight-qc flag."""
-        args = parse_args(["execute", "feature", "--skip-preflight-qc"])
-        assert args.skip_preflight_qc is True
-
-    def test_skip_preflight_qc_flag_defaults_false(self) -> None:
-        """parse_args() defaults skip_preflight_qc to False."""
-        args = parse_args(["execute", "feature"])
-        assert args.skip_preflight_qc is False
-
-
-class TestPhaseEndQC:
-    """Tests for phase-end QC behavior in the CLI."""
-
-    def test_phase_end_expectations_passed_to_qc_runner(
-        self, monkeypatch: "MonkeyPatch"
-    ) -> None:
-        """
-        CLI should pass resolved expectations to phase-end QC.
-
-        Purpose:
-            Ensure phase completion forwards plan expectations into run_full().
-        """
-        from scripts.dev_tools.atomic_executor import cli
-        from scripts.dev_tools.atomic_executor.plan_parser import PlanModel, PlanTask
-
-        plan_task = PlanTask(
-            task_id="P1-T1",
-            phase=1,
-            task_num=1,
-            title="pytest tests/bugs/2026/test_issue_98.py::test_expected_fail",
-            checked=True,
-            line_index=0,
-            expect_fail=True,
-            test_ref="tests/bugs/2026/test_issue_98.py::test_expected_fail",
-        )
-        plan = PlanModel(tasks=[plan_task], phases=[1])
-
-        parser = Mock()
-        parser.next_unchecked_task.return_value = plan_task
-        parser.find_task_by_id.return_value = plan_task
-        parser.phase_complete.return_value = True
-        parser.preflight_validate.return_value = None
-        parser.parse.return_value = plan
-
-        resolver = Mock()
-        resolver.resolve.return_value = (Mock(), Path.cwd())
-        qc_runner = Mock()
-
-        monkeypatch.setattr(cli, "PlanParser", lambda _path: parser)
-        monkeypatch.setattr(cli, "FeatureResolver", Mock(return_value=resolver))
-        monkeypatch.setattr(cli, "QCRunner", Mock(return_value=qc_runner))
-        monkeypatch.setattr(cli, "resolve_workspace", lambda _workspace: Path.cwd())
-        monkeypatch.setattr(
-            cli,
-            "resolve_feature_plan",
-            lambda _dir: Mock(path=Path("docs/features/active/README.md")),
-        )
-        monkeypatch.setattr(cli, "refuse_protected_branch", lambda _workspace: None)
-        monkeypatch.setattr(cli, "execute_one_task", lambda **_kwargs: 0)
-
-        exit_code = cli.main(["execute", "feature", "--skip-preflight-qc"])
-
-        assert exit_code == 0
-        qc_runner.run_full.assert_called_once()
-        _, kwargs = qc_runner.run_full.call_args
-        assert "expectations" in kwargs
-        assert (
-            "tests/bugs/2026/test_issue_98.py::test_expected_fail"
-            in kwargs["expectations"].expected_fail_refs
-        )
-
-    def test_preflight_expected_fail_allows_known_failures(
-        self,
-        monkeypatch: "MonkeyPatch",
-    ) -> None:
-        """
-        Preflight QC should allow failures covered by expected-fail refs.
-
-        Purpose:
-            Ensure expected-fail refs suppress baseline fix behavior for pytest.
-        """
-        from scripts.dev_tools.atomic_executor.cli import (
-            _run_preflight_qc_with_capture,
-        )
-        from scripts.dev_tools.atomic_executor.pytest_expectations import (
-            ResolvedTestExpectations,
-        )
-
-        expectations = ResolvedTestExpectations(
-            expected_fail_refs={"tests/bugs/2026/test_issue_98.py::test_expected_fail"},
-            expected_pass_refs=set(),
-            expected_fail_jest_refs=set(),
-            expected_pass_jest_refs=set(),
-            missing_test_refs=[],
-        )
-
-        def mock_run(
-            *args: object, **kwargs: object
-        ) -> subprocess.CompletedProcess[str]:
-            """Return a failing pytest run for the expected-fail nodeid."""
-            cmd = args[0]
-            if isinstance(cmd, list) and "pytest" in cmd and "--collect-only" in cmd:
-                return subprocess.CompletedProcess(
-                    args=cmd, returncode=0, stdout="collected", stderr=""
-                )
-            if isinstance(cmd, list) and "pytest" in cmd:
-                output = (
-                    "FAILED tests/bugs/2026/test_issue_98.py::"
-                    "test_expected_fail - AssertionError"
-                )
-                return subprocess.CompletedProcess(
-                    args=cmd, returncode=1, stdout=output, stderr=""
-                )
-            return subprocess.CompletedProcess(
-                args=cmd, returncode=0, stdout="OK", stderr=""
-            )
-
-        monkeypatch.setattr("subprocess.run", mock_run)
-
-        result = _run_preflight_qc_with_capture(Path.cwd(), expectations=expectations)
-
-        assert result.success is True
-        assert result.failed_step is None
-        assert "expected pytest failures" in result.output.lower()
-
-    def test_preflight_expected_fail_allows_known_jest_failures(
-        self,
-        monkeypatch: "MonkeyPatch",
-    ) -> None:
-        """
-        Preflight QC should allow failures covered by expected Jest refs.
-
-        Purpose:
-            Ensure expected-fail refs suppress baseline fix behavior for Jest.
-        """
-        from scripts.dev_tools.atomic_executor.cli import (
-            _run_preflight_qc_with_capture,
-        )
-        from scripts.dev_tools.atomic_executor.pytest_expectations import (
-            ResolvedTestExpectations,
-        )
-        from scripts.dev_tools.atomic_executor.qc_toolchain import QCToolchain
-
-        expectations = ResolvedTestExpectations(
-            expected_fail_refs=set(),
-            expected_pass_refs=set(),
-            missing_test_refs=[],
-            expected_fail_jest_refs={
-                "tests/unit/task-execution-spec.test.ts::"
-                "getTaskExecutionSpec returns QC black"
-            },
-            expected_pass_jest_refs=set(),
-        )
-
-        def mock_run(
-            *args: object, **kwargs: object
-        ) -> subprocess.CompletedProcess[str]:
-            """Return a failing npm test run for the expected-fail ref."""
-            cmd = args[0]
-            if isinstance(cmd, list) and cmd[:3] == ["npm", "run", "test:unit"]:
-                output = "\n".join(
-                    [
-                        "FAIL tests/unit/task-execution-spec.test.ts",
-                        "  \u25cf getTaskExecutionSpec returns QC black",
-                    ]
-                )
-                return subprocess.CompletedProcess(
-                    args=cmd, returncode=1, stdout=output, stderr=""
-                )
-            return subprocess.CompletedProcess(
-                args=cmd, returncode=0, stdout="OK", stderr=""
-            )
-
-        monkeypatch.setattr("subprocess.run", mock_run)
-
-        result = _run_preflight_qc_with_capture(
-            Path.cwd(), expectations=expectations, toolchain=QCToolchain.TYPESCRIPT
-        )
-
-        assert result.success is True
-        assert result.failed_step is None
-
-    def test_preflight_expected_pass_wins(
-        self,
-        monkeypatch: "MonkeyPatch",
-    ) -> None:
-        """
-        Expected-pass should override expected-fail for the same ref.
-
-        Purpose:
-            Ensure expected-pass entries cause the pytest gate to fail.
-        """
-        from scripts.dev_tools.atomic_executor.cli import (
-            _run_preflight_qc_with_capture,
-        )
-        from scripts.dev_tools.atomic_executor.pytest_expectations import (
-            ResolvedTestExpectations,
-        )
-
-        expectations = ResolvedTestExpectations(
-            expected_fail_refs={"tests/bugs/2026/test_issue_98.py::test_expected_fail"},
-            expected_pass_refs={"tests/bugs/2026/test_issue_98.py::test_expected_fail"},
-            expected_fail_jest_refs=set(),
-            expected_pass_jest_refs=set(),
-            missing_test_refs=[],
-        )
-
-        def mock_run(
-            *args: object, **kwargs: object
-        ) -> subprocess.CompletedProcess[str]:
-            """Return a failing pytest run for the expected-pass nodeid."""
-            cmd = args[0]
-            if isinstance(cmd, list) and "pytest" in cmd and "--collect-only" in cmd:
-                return subprocess.CompletedProcess(
-                    args=cmd, returncode=0, stdout="collected", stderr=""
-                )
-            if isinstance(cmd, list) and "pytest" in cmd:
-                output = (
-                    "FAILED tests/bugs/2026/test_issue_98.py::"
-                    "test_expected_fail - AssertionError"
-                )
-                return subprocess.CompletedProcess(
-                    args=cmd, returncode=1, stdout=output, stderr=""
-                )
-            return subprocess.CompletedProcess(
-                args=cmd, returncode=0, stdout="OK", stderr=""
-            )
-
-        monkeypatch.setattr("subprocess.run", mock_run)
-
-        result = _run_preflight_qc_with_capture(Path.cwd(), expectations=expectations)
-
-        assert result.success is False
-        assert result.failed_step == "pytest"
-        assert "expected-pass" in result.output.lower()
-
-    def test_preflight_missing_test_ref(
-        self,
-        monkeypatch: "MonkeyPatch",
-    ) -> None:
-        """
-        Missing test refs should fail preflight before running pytest.
-
-        Purpose:
-            Provide actionable feedback when expectation tasks lack test refs.
-        """
-        from scripts.dev_tools.atomic_executor.cli import (
-            _run_preflight_qc_with_capture,
-        )
-        from scripts.dev_tools.atomic_executor.pytest_expectations import (
-            ResolvedTestExpectations,
-        )
-
-        expectations = ResolvedTestExpectations(
-            expected_fail_refs=set(),
-            expected_pass_refs=set(),
-            expected_fail_jest_refs=set(),
-            expected_pass_jest_refs=set(),
-            missing_test_refs=["P1-T1"],
-        )
-
-        def mock_run(
-            *args: object, **kwargs: object
-        ) -> subprocess.CompletedProcess[str]:
-            """Return success for non-pytest steps."""
-            cmd = args[0]
-            return subprocess.CompletedProcess(
-                args=cmd, returncode=0, stdout="OK", stderr=""
-            )
-
-        monkeypatch.setattr("subprocess.run", mock_run)
-
-        result = _run_preflight_qc_with_capture(Path.cwd(), expectations=expectations)
-
-        assert result.success is False
-        assert result.failed_step == "pytest-collect"
-        assert "missing test reference" in result.output.lower()
-
-
-class TestRunCopilot:
-    """Tests for run_copilot() function."""
-
-    def test_run_copilot_raises_when_executable_not_found(
-        self, mem_path: Path, monkeypatch: "MonkeyPatch"
-    ) -> None:
-        """run_copilot() raises FileNotFoundError when copilot not found."""
-        from scripts.dev_tools.atomic_executor.cli import run_copilot
-
-        monkeypatch.setenv("XDG_CONFIG_HOME", str(mem_path / "config-root"))
-        # Set PATH to empty so no copilot executable can be found
-        monkeypatch.setenv("PATH", "")
-
-        log_file = mem_path / "test.log"
-
-        with pytest.raises(
-            FileNotFoundError,
-            match=r"Required executable not found on PATH: copilot\b",
-        ):
-            run_copilot(
-                workspace=mem_path,
-                prompt_text="test prompt",
-                log_file=log_file,
-                task_id="P1-T1",
-                preferred_model=None,
-                run_id="2026-01-07_000000",
-            )
-
-    def test_run_copilot_rejects_vscode_shim(
-        self, mem_path: Path, monkeypatch: "MonkeyPatch"
-    ) -> None:
-        """run_copilot() skips VS Code shim and finds no other copilot."""
-        from scripts.dev_tools.atomic_executor.cli import run_copilot
-
-        monkeypatch.setenv("XDG_CONFIG_HOME", str(mem_path / "config-root"))
-        # Create shim directory structure that matches the detection pattern
-        # Pattern: .../Code/User/globalStorage/github.copilot-chat/copilotCli/
-        # Use nested dirs to ensure we hit the pattern matching logic
-        shim_dir = (
-            mem_path
-            / "Code"
-            / "User"
-            / "globalStorage"
-            / "github.copilot-chat"
-            / "copilotCli"
-        )
-        shim_dir.mkdir(parents=True)
-
-        # Create executable shim
-        # In cli.py, it checks specifically for copilot.exe, copilot.bat, copilot
-        fake_shim = shim_dir / "copilot"
-        fake_shim.touch()
-        fake_shim.chmod(0o755)
-
-        # Update PATH to point to this directory
-        monkeypatch.setenv("PATH", str(shim_dir))
-
-        log_file = mem_path / "test.log"
-
-        with pytest.raises(
-            FileNotFoundError,
-            match=r"Required executable not found on PATH: copilot\b",
-        ):
-            run_copilot(
-                workspace=mem_path,
-                prompt_text="test prompt",
-                log_file=log_file,
-                task_id="P1-T1",
-                preferred_model=None,
-                run_id="2026-01-07_000000",
-            )
-
-    def test_run_copilot_rejects_vscode_shim_remote_paths(
-        self, mem_path: Path, monkeypatch: "MonkeyPatch"
-    ) -> None:
-        """run_copilot() rejects the VS Code Remote/Devcontainer shim path."""
-        from scripts.dev_tools.atomic_executor.cli import run_copilot
-
-        monkeypatch.setenv("XDG_CONFIG_HOME", str(mem_path / "config-root"))
-        # VS Code Remote (including devcontainers) stores its shim under a Linux
-        # path, e.g. ~/.vscode-server/data/User/globalStorage/github.copilot-chat/
-        # copilotCli/. If we accidentally execute this shim, it can block waiting
-        # for interactive install/auth and appear as a hang.
-        shim_dir = (
-            mem_path
-            / ".vscode-server"
-            / "data"
-            / "User"
-            / "globalStorage"
-            / "github.copilot-chat"
-            / "copilotCli"
-        )
-        shim_dir.mkdir(parents=True)
-
-        fake_shim = shim_dir / "copilot"
-        fake_shim.touch()
-        fake_shim.chmod(0o755)
-
-        # Ensure the shim is the ONLY PATH entry.
-        monkeypatch.setenv("PATH", str(shim_dir))
-
-        # Guard: if the implementation tries to execute the shim, fail the test.
-        def _should_not_invoke_popen(*args: object, **kwargs: object) -> None:
-            pytest.fail("run_copilot attempted to execute a VS Code shim")
-
-        monkeypatch.setattr("subprocess.Popen", _should_not_invoke_popen)
-
-        log_file = mem_path / "test.log"
-
-        with pytest.raises(
-            FileNotFoundError,
-            match=r"Required executable not found on PATH: copilot\b",
-        ):
-            run_copilot(
-                workspace=mem_path,
-                prompt_text="test prompt",
-                log_file=log_file,
-                task_id="P1-T1",
-                preferred_model=None,
-                run_id="2026-01-07_000000",
-            )
-
-    def test_run_copilot_creates_log_directory(
-        self, mem_path: Path, monkeypatch: "MonkeyPatch"
-    ) -> None:
-        """run_copilot() creates log directory if missing."""
-        from scripts.dev_tools.atomic_executor.cli import run_copilot
-
-        monkeypatch.setenv("XDG_CONFIG_HOME", str(mem_path / "config-root"))
-        # Setup fake copilot on PATH
-        bin_dir = mem_path / "bin"
-        bin_dir.mkdir()
-        copilot_exe = bin_dir / "copilot"
-        copilot_exe.touch()
-        copilot_exe.chmod(0o755)
-        path = os.environ.get("PATH", "")
-        monkeypatch.setenv("PATH", f"{str(bin_dir)}{os.pathsep}{path}")
-
-        class MockStdout:
-            def read(self, size: int = -1) -> bytes:
-                return b""
-
-        class MockPopen:
-            def __init__(
-                self, argv: list[str], *args: object, **kwargs: object
-            ) -> None:
-                self.stdout = MockStdout()
-                self.returncode = 0
-
-            def poll(self) -> int:
-                return 0
-
-            def wait(self) -> int:
-                return 0
-
-        monkeypatch.setattr("subprocess.Popen", MockPopen)
-
-        # Also mock run purely to avoid confusion, though unused by run_copilot directly
-        def mock_run(*args: object, **kwargs: object) -> Mock:
-            return Mock(returncode=0)
-
-        monkeypatch.setattr("subprocess.run", mock_run)
-
-        log_dir = mem_path / "nested" / "log" / "dir"
-        log_file = log_dir / "test.log"
-
-        run_copilot(
-            workspace=mem_path,
-            prompt_text="test prompt",
-            log_file=log_file,
-            task_id="P1-T1",
-            preferred_model=None,
-            run_id="2026-01-07_000000",
-        )
-
-        assert log_dir.exists()
-        assert log_file.exists()
-
-    def test_run_copilot_prefers_cmd_wrapper_over_bare_executable_name(
-        self, mem_path: Path, monkeypatch: "MonkeyPatch"
-    ) -> None:
-        """run_copilot() prefers copilot.cmd over a bare 'copilot' file on Windows."""
-
-        from scripts.dev_tools.atomic_executor.cli import run_copilot
-
-        monkeypatch.setenv("XDG_CONFIG_HOME", str(mem_path / "config-root"))
-
-        # Arrange a PATH entry with both:
-        # - copilot (often a POSIX shim from npm, not executable by CreateProcess)
-        # - copilot.cmd (Windows-friendly wrapper)
-        fake_bin = mem_path / "bin"
-        fake_bin.mkdir()
-        (fake_bin / "copilot").write_text("#!/usr/bin/env node\n", encoding="utf-8")
-        (fake_bin / "copilot.cmd").write_text("@echo fake copilot\n", encoding="utf-8")
-        monkeypatch.setenv("PATH", str(fake_bin))
-
-        captured_argv: list[str] = []
-
-        class MockStdout:
-            def read(self, size: int = -1) -> bytes:
-                return b""
-
-        class MockPopen:
-            def __init__(
-                self, argv: list[str], *args: object, **kwargs: object
-            ) -> None:
-                captured_argv.extend(argv)
-                self.stdout = MockStdout()
-                self.returncode = 0
-
-            def poll(self) -> int:
-                return 0
-
-            def wait(self) -> int:
-                return 0
-
-        monkeypatch.setattr("subprocess.Popen", MockPopen)
-
-        log_file = mem_path / "test.log"
-
-        run_copilot(
-            workspace=mem_path,
-            prompt_text="test prompt",
-            log_file=log_file,
-            task_id="P1-T1",
-            preferred_model=None,
-            run_id="2026-01-07_000000",
-        )
-
-        assert captured_argv
-        assert Path(captured_argv[0]).name == "copilot.cmd"
-
-    def test_run_copilot_invokes_with_correct_arguments(
-        self, mem_path: Path, monkeypatch: "MonkeyPatch"
-    ) -> None:
-        """run_copilot() invokes copilot with correct arguments."""
-        from scripts.dev_tools.atomic_executor.cli import run_copilot
-
-        monkeypatch.setenv("XDG_CONFIG_HOME", str(mem_path / "config-root"))
-        # Create a fake copilot executable on PATH
-        fake_bin = mem_path / "bin"
-        fake_bin.mkdir()
-        fake_copilot = fake_bin / "copilot.exe"
-        fake_copilot.write_text("@echo fake copilot")
-        monkeypatch.setenv("PATH", str(fake_bin))
-
-        captured_argv: list[str] = []
-        captured_stdin: str | None = None
-        captured_stdin_was_provided = False
-
-        class MockStdout:
-            def read(self, size: int = -1) -> bytes:
-                return b""
-
-        class MockPopen:
-            def __init__(
-                self, argv: list[str], *args: object, **kwargs: object
-            ) -> None:
-                nonlocal captured_stdin
-                nonlocal captured_stdin_was_provided
-                captured_argv.extend(argv)
-                # Capture stdin content if provided
-                stdin_arg = kwargs.get("stdin")
-                captured_stdin_was_provided = (
-                    "stdin" in kwargs and stdin_arg is not None
-                )
-                if stdin_arg and hasattr(stdin_arg, "read"):
-                    # Type narrowing: we know it has a read method and returns bytes
-                    raw_content = stdin_arg.read()  # type: ignore[union-attr]
-                    # Explicitly cast to bytes for type checker
-                    content_bytes: bytes = (
-                        raw_content if isinstance(raw_content, bytes) else b""
-                    )
-                    captured_stdin = content_bytes.decode("utf-8")
-                self.stdout = MockStdout()
-                self.returncode = 0
-
-            def poll(self) -> int:
-                return 0
-
-            def wait(self) -> int:
-                return 0
-
-        monkeypatch.setattr("subprocess.Popen", MockPopen)
-
-        log_file = mem_path / "test.log"
-
-        run_copilot(
-            workspace=mem_path,
-            prompt_text="test prompt",
-            log_file=log_file,
-            task_id="P1-T1",
-            preferred_model="gpt-5.1-codex-max",
-            run_id="2026-01-07_000000",
-        )
-
-        expected_prompt_file = (
-            log_file.parent / "prompts" / "prompt_2026-01-07_000000_P1-T1.md"
-        )
-        assert expected_prompt_file.exists()
-        assert expected_prompt_file.read_text(encoding="utf-8") == "test prompt"
-
-        assert captured_argv[0] == str(fake_copilot)
-        assert "--agent" in captured_argv
-        agent_idx = captured_argv.index("--agent")
-        assert captured_argv[agent_idx + 1] == "atomic_executor"
-        assert "--model" in captured_argv
-        assert "gpt-5.1-codex-max" in captured_argv
-        assert "--session-path" not in captured_argv
-        # Prompt must be passed via programmatic mode (-p/--prompt) with an @path
-        # reference to the on-disk prompt file.
-        assert "-p" in captured_argv or "--prompt" in captured_argv
-        prompt_flag = "-p" if "-p" in captured_argv else "--prompt"
-        prompt_idx = captured_argv.index(prompt_flag)
-        prompt_value = captured_argv[prompt_idx + 1]
-        assert f"@{expected_prompt_file}" in prompt_value
-
-        # Prompt must NOT be provided via stdin.
-        assert captured_stdin_was_provided is False
-        assert captured_stdin is None
-        assert "--share" in captured_argv
-        assert "--allow-tool" in captured_argv
-        assert "write" in captured_argv
-
-        # Headless-friendly defaults must include broad shell and path permissions.
-        assert "--allow-tool" in captured_argv
-        assert "shell" in captured_argv
-        assert "--allow-all-paths" in captured_argv
-
-        # Tool approvals required for headless QC must remain present.
-        assert "shell(poetry)" in captured_argv
-        assert "shell(git)" in captured_argv
-
-    def test_run_copilot_normalizes_gpt_5_2_codex_display_name(
-        self, mem_path: Path, monkeypatch: "MonkeyPatch"
-    ) -> None:
-        """run_copilot() normalizes GPT-5.2 Codex display names for Copilot CLI."""
-
-        from scripts.dev_tools.atomic_executor.cli import run_copilot
-
-        monkeypatch.setenv("XDG_CONFIG_HOME", str(mem_path / "config-root"))
-        # Create a fake copilot executable on PATH.
-        fake_bin = mem_path / "bin"
-        fake_bin.mkdir()
-        fake_copilot = fake_bin / "copilot.exe"
-        fake_copilot.write_text("@echo fake copilot")
-        monkeypatch.setenv("PATH", str(fake_bin))
-
-        captured_argv: list[str] = []
-
-        class MockStdout:
-            def read(self, size: int = -1) -> bytes:
-                return b""
-
-        class MockPopen:
-            def __init__(
-                self, argv: list[str], *args: object, **kwargs: object
-            ) -> None:
-                captured_argv.extend(argv)
-                self.stdout = MockStdout()
-                self.returncode = 0
-
-            def poll(self) -> int:
-                return 0
-
-            def wait(self) -> int:
-                return 0
-
-        monkeypatch.setattr("subprocess.Popen", MockPopen)
-
-        log_file = mem_path / "test.log"
-
-        # VS Code tasks sometimes pass this exact display-style model string.
-        # The executor should normalize it to the Copilot CLI model key.
-        run_copilot(
-            workspace=mem_path,
-            prompt_text="test prompt",
-            log_file=log_file,
-            task_id="P1-T1",
-            preferred_model="GPT-5.2-Codex",
-            run_id="2026-01-07_000000",
-        )
-
-        assert "--model" in captured_argv
-        model_idx = captured_argv.index("--model")
-        assert captured_argv[model_idx + 1] == "gpt-5.2-codex"
-
-    def test_run_copilot_permission_denied_fails_fast_with_actionable_error(
-        self, mem_path: Path, monkeypatch: "MonkeyPatch"
-    ) -> None:
-        """run_copilot() raises promptly when Copilot reports a permission denial."""
-
-        from scripts.dev_tools.atomic_executor.cli import run_copilot
-
-        monkeypatch.setenv("XDG_CONFIG_HOME", str(mem_path / "config-root"))
-        # Create a fake copilot executable on PATH.
-        fake_bin = mem_path / "bin"
-        fake_bin.mkdir()
-        fake_copilot = fake_bin / "copilot.exe"
-        fake_copilot.write_text("@echo fake copilot")
-        monkeypatch.setenv("PATH", str(fake_bin))
-
-        captured_argv: list[str] = []
-
-        permission_denied = (
-            "Permission denied and could not request permission from user"
-        )
-
-        class MockStdout:
-            def __init__(self) -> None:
-                self._chunks = [permission_denied.encode("utf-8"), b""]
-
-            def read(self, size: int = -1) -> bytes:
-                return self._chunks.pop(0) if self._chunks else b""
-
-            def read1(self, size: int = -1) -> bytes:
-                return self.read(size)
-
-        class MockPopen:
-            def __init__(
-                self, argv: list[str], *args: object, **kwargs: object
-            ) -> None:
-                captured_argv.extend(argv)
-                self.stdout = MockStdout()
-                self.returncode = 1
-
-            def poll(self) -> int:
-                return 1
-
-            def wait(self, timeout: float | None = None) -> int:
-                return 1
-
-        monkeypatch.setattr("subprocess.Popen", MockPopen)
-
-        log_file = mem_path / "test.log"
-
-        with pytest.raises(RuntimeError) as exc_info:
-            run_copilot(
-                workspace=mem_path,
-                prompt_text="test prompt",
-                log_file=log_file,
-                task_id="P1-T1",
-                preferred_model="gpt-5.1-codex-max",
-                run_id="2026-01-07_000000",
-            )
-
-        error_text = str(exc_info.value)
-        assert permission_denied in error_text
-        assert "copilot" in error_text.lower()
-        assert "-p" in error_text or "--prompt" in error_text
-        assert "--allow-tool" in error_text
-        assert "write" in error_text
-        assert "shell(poetry)" in error_text
-        assert "shell(git)" in error_text
-        assert "--allow-all-paths" in error_text
-
-        # Sanity check the argv we actually attempted to run.
-        assert captured_argv
-        assert captured_argv[0] == str(fake_copilot)
-
-    def test_run_copilot_reuses_session_when_requested(
-        self, mem_path: Path, monkeypatch: "MonkeyPatch"
-    ) -> None:
-        """run_copilot() adds --continue when resume_session=True."""
-        from scripts.dev_tools.atomic_executor.cli import run_copilot
-
-        monkeypatch.setenv("XDG_CONFIG_HOME", str(mem_path / "config-root"))
-        fake_bin = mem_path / "bin"
-        fake_bin.mkdir()
-        fake_copilot = fake_bin / "copilot"
-        fake_copilot.write_text("#!/bin/sh\necho copilot")
-        monkeypatch.setenv("PATH", str(fake_bin))
-
-        captured_argv: list[str] = []
-
-        class MockStdout:
-            def read(self, size: int = -1) -> bytes:
-                return b""
-
-        class MockPopen:
-            def __init__(
-                self, argv: list[str], *args: object, **kwargs: object
-            ) -> None:
-                captured_argv.extend(argv)
-                self.stdout = MockStdout()
-                self.returncode = 0
-
-            def poll(self) -> int:
-                return 0
-
-            def wait(self) -> int:
-                return 0
-
-        monkeypatch.setattr("subprocess.Popen", MockPopen)
-
-        log_file = mem_path / "log" / "test.log"
-
-        run_copilot(
-            workspace=mem_path,
-            prompt_text="retry prompt",
-            log_file=log_file,
-            task_id="P1-T1",
-            preferred_model=None,
-            run_id="2026-01-07_000000",
-            resume_session=True,
-        )
-
-        # resume_session=True should add --continue flag
-        assert "--continue" in captured_argv
-
-    def test_run_copilot_trusts_workspace_in_config(
-        self, mem_path: Path, monkeypatch: "MonkeyPatch"
-    ) -> None:
-        """run_copilot() writes workspace to trusted_folders when enabled."""
-        from scripts.dev_tools.atomic_executor.cli import run_copilot
-
-        # Route Copilot config to a temp directory for isolation.
-        config_root = mem_path / "config-root"
-        monkeypatch.setenv("XDG_CONFIG_HOME", str(config_root))
-
-        fake_bin = mem_path / "bin"
-        fake_bin.mkdir()
-        fake_copilot = fake_bin / "copilot"
-        fake_copilot.write_text("#!/bin/sh\necho copilot")
-        monkeypatch.setenv("PATH", str(fake_bin))
-
-        class MockStdout:
-            def read(self, size: int = -1) -> bytes:
-                return b""
-
-        class MockPopen:
-            def __init__(
-                self, argv: list[str], *args: object, **kwargs: object
-            ) -> None:
-                self.stdout = MockStdout()
-                self.returncode = 0
-
-            def poll(self) -> int:
-                return 0
-
-            def wait(self) -> int:
-                return 0
-
-        monkeypatch.setattr("subprocess.Popen", MockPopen)
-
-        log_file = mem_path / "log" / "test.log"
-
-        run_copilot(
-            workspace=mem_path,
-            prompt_text="trust prompt",
-            log_file=log_file,
-            task_id="P1-T1",
-            preferred_model=None,
-            run_id="2026-01-07_000000",
-            allow_all_paths=True,
-            allow_all_urls=False,
-            allow_shell=True,
-            trust_workspace=True,
-        )
-
-        config_file = config_root / "copilot" / "config.json"
-        assert config_file.exists()
-        config_text = config_file.read_text(encoding="utf-8")
-        import json
-
-        config_data = json.loads(config_text)
-        trusted_folders = config_data.get("trusted_folders", [])
-        assert str(mem_path.resolve()) in trusted_folders
-
-    def test_run_copilot_times_out_when_cli_is_idle(
-        self, mem_path: Path, monkeypatch: "MonkeyPatch"
-    ) -> None:
-        """run_copilot() terminates when Copilot CLI produces no output."""
-        from scripts.dev_tools.atomic_executor.cli import run_copilot
-
-        monkeypatch.setenv("XDG_CONFIG_HOME", str(mem_path / "config-root"))
-        bin_dir = mem_path / "bin"
-        bin_dir.mkdir()
-        copilot_exe = bin_dir / "copilot"
-        copilot_exe.write_text("#!/bin/sh\nexit 0")
-        copilot_exe.chmod(0o755)
-
-        path = os.environ.get("PATH", "")
-        monkeypatch.setenv("PATH", f"{str(bin_dir)}{os.pathsep}{path}")
-        monkeypatch.setenv("ATOMIC_EXECUTOR_COPILOT_IDLE_TIMEOUT_SECONDS", "0.2")
-
-        stdout_r, stdout_w = os.pipe()
-        os.close(stdout_w)
-        stdout_stream = os.fdopen(stdout_r, "rb", buffering=0)
-
-        hung_process_holder: dict[str, object] = {}
-
-        class HungProcess:
-            def __init__(self, argv: list[str]) -> None:
-                self.args = argv
-                self.stdout = stdout_stream
-                self.returncode: int | None = None
-                self.killed = False
-
-            def poll(self) -> int | None:
-                return self.returncode
-
-            def wait(self, timeout: float | None = None) -> int | None:
-                return self.returncode
-
-            def kill(self) -> None:
-                self.killed = True
-                self.returncode = -9
-
-        def fake_popen(argv: list[str], *args: object, **kwargs: object) -> HungProcess:
-            proc = HungProcess(argv)
-            hung_process_holder["proc"] = proc
-            return proc
-
-        monkeypatch.setattr("subprocess.Popen", fake_popen)
-
-        log_file = mem_path / "log" / "test.log"
-
-        with pytest.raises(TimeoutError):
-            run_copilot(
-                workspace=mem_path,
-                prompt_text="idle prompt",
-                log_file=log_file,
-                task_id="P1-T1",
-                preferred_model=None,
-                run_id="2026-01-07_000000",
-            )
-
-        proc = hung_process_holder.get("proc")
-        assert proc is not None
-        assert getattr(proc, "killed", False) is True
-        stdout_stream.close()
