+++ tests/scripts/dev_tools/test_collect_pr_context.py
@@ -15,7 +15,6 @@ from scripts.dev_tools.pr_context.collector import (
     PullRequestDetails,
     build_close_candidates_section,
     build_pr_context,
-    collect_and_write,
     convert_numstat,
     extension_summary,
     extract_issue_references,
@@ -23,17 +22,7 @@ from scripts.dev_tools.pr_context.collector import (
     find_user_story_link,
     format_diff_path,
     gather_feature_excerpts,
-    main,
-    parse_args,
     select_default_base,
-    write_output,
-)
-from scripts.dev_tools.pr_context.models import FeatureDocExcerpt, PRContextResult
-from scripts.dev_tools.pr_context.summary_helpers import (
-    extract_digest_bullets as _extract_digest_bullets,
-)
-from scripts.dev_tools.pr_context.summary_helpers import (
-    is_scoping_doc as _is_scoping_doc,
 )
 from scripts.dev_tools.pr_context.summary_helpers import (
     issue_appendix as _issue_appendix,
@@ -49,12 +38,6 @@ from scripts.dev_tools.pr_context.summary_helpers import (
 )
 from scripts.dev_tools.pr_context.summary_helpers import (
     parse_numstat_detailed as _parse_numstat_detailed,
-)
-from scripts.dev_tools.pr_context.summary_helpers import (
-    pr_appendix as _pr_appendix,
-)
-from scripts.dev_tools.pr_context.summary_helpers import (
-    pr_digest as _pr_digest,
 )
 from scripts.dev_tools.pr_context.summary_helpers import (
     scoping_doc_changes as _scoping_doc_changes,
@@ -457,779 +440,3 @@ def test_scoping_doc_changes_marks_non_material_link_only():
     )
     assert changes
     assert changes[0][1] is False
-
-
-def test_pr_digest_and_appendix_cover_headings_and_files():
-    pr = PullRequestDetails(
-        number="#21",
-        title="Improve docs",
-        state="open",
-        author="lee",
-        base_ref="main",
-        head_ref="feature/docs",
-        created_at="2024-01-01",
-        updated_at="2024-01-02",
-        merged_at=None,
-        labels=["docs"],
-        assignees=["lee"],
-        body="## Why\n- clarify usage\n",
-        closing_issues=["#8"],
-        files_changed=["a.md", "b.md", "c.md", "d.md"],
-    )
-    digest = _pr_digest(pr)
-    appendix = _pr_appendix(pr)
-    assert "Why: clarify usage" in digest
-    assert "Files (first 25)" in appendix
-    assert "closes" not in appendix.lower()
-
-
-def test_issue_appendix_includes_user_story_block():
-    issue = IssueDetails(
-        number="#12",
-        title="Story reference",
-        state="open",
-        labels=[],
-        assignees=[],
-        author="alex",
-        created_at="2024-03-01",
-        updated_at="2024-03-02",
-        body="Context\n" + "\n".join(f"line {i}" for i in range(10)),
-        comments=["note"],
-        user_story_path="docs/story/user-story.md",
-        user_story_content="Story content line 1\nline 2",
-    )
-    appendix = _issue_appendix(issue)
-    assert "User story (docs/story/user-story.md)" in appendix
-    assert "Story content line" in appendix
-
-
-def test_parse_numstat_detailed_handles_non_numeric_entries():
-    adds, dels, mapping = _parse_numstat_detailed(
-        "-\t-\tfirst.txt\nnotnum\t3\tsecond.txt\n"
-    )
-    assert adds == 0
-    assert dels == 3
-    assert mapping["first.txt"] == (0, 0)
-    assert mapping["second.txt"] == (0, 3)
-
-
-def test_is_scoping_doc_identifies_feature_files():
-    assert _is_scoping_doc("docs/features/active/feat/spec.md")
-    assert _is_scoping_doc("docs/features/active/feat/plan.md")
-    assert _is_scoping_doc("docs/features/active/feat/bug-remediation-plan.md")
-    assert _is_scoping_doc("docs/features/active/feat/user-story.md")
-    assert not _is_scoping_doc("docs/features/ideas/idea.md")
-    assert not _is_scoping_doc("src/main.py")
-
-
-def test_collect_and_write_uses_feature_refs_and_scoping(
-    monkeypatch: pytest.MonkeyPatch, mem_path: Path
-) -> None:
-    captured: list[tuple[Path, str]] = []
-
-    def fake_write_output(text: str, out_path: Path, append: bool) -> None:
-        captured.append((out_path, text))
-
-    monkeypatch.setattr(
-        "scripts.dev_tools.pr_context.collector.write_output", fake_write_output
-    )
-
-    class FakeGit:
-        def __init__(self, *args: object, **kwargs: object) -> None:
-            self._root = Path(__file__).resolve().parents[3]
-
-        def resolve_root(self) -> Path:
-            return self._root
-
-        def branch_name(self) -> str:
-            return "feature/ABC-10"
-
-        def diff_range(self, args: Sequence[str]) -> str:
-            path = "docs/features/active/fix-all-script/spec.md"
-            if "--name-status" in args:
-                return f"M\t{path}"
-            if "--numstat" in args:
-                return f"20\t0\t{path}"
-            if "--unified=0" in args:
-                return f"+++ b/{path}\n+## Acceptance Criteria\n+New criteria"
-            return ""
-
-        def rev_parse(self, ref: str) -> str:
-            return f"{ref}-sha"
-
-    class FakeGh:
-        status_message = "ok"
-
-        def __init__(self, *args: object, **kwargs: object) -> None:
-            return None
-
-        def ensure_available(self) -> None:
-            return None
-
-        def current_pr(self) -> PullRequestDetails | None:
-            return None
-
-        def classify_entity(self, number: str) -> str | None:
-            return "issue" if number in {"1", "ABC-10"} else "pull"
-
-        def issue_details(self, number: str) -> IssueDetails:
-            return IssueDetails(
-                number=f"#{number}",
-                title="Issue",
-                state="open",
-                labels=[],
-                assignees=[],
-                author="alex",
-                created_at="2024-01-01",
-                updated_at="2024-01-02",
-                body="Body",
-                comments=[],
-            )
-
-        def pr_details(self, number: str) -> PullRequestDetails:
-            return PullRequestDetails(
-                number=f"#{number}",
-                title="PR",
-                state="open",
-                author="alex",
-                base_ref="main",
-                head_ref="feature",
-                created_at="2024-01-01",
-                updated_at="2024-01-02",
-                merged_at=None,
-                labels=[],
-                assignees=[],
-                body="PR body",
-                closing_issues=[],
-                files_changed=["file.py"],
-            )
-
-        def ci_status(self, head_sha: str) -> tuple[str | None, list[str]]:
-            return "success", []
-
-    feature_calls: list[list[str]] = []
-
-    def fake_build_pr_context(
-        *,
-        git: GitClient,
-        gh: FakeGh,
-        base_ref: str | None,
-        head_ref: str | None,
-        include_untracked: bool,
-        feature_issue_refs: Sequence[str] | None = None,
-        current_pr: PullRequestDetails | None = None,
-        gh_available: bool | None = None,
-    ) -> PRContextResult:
-        feature_calls.append(list(feature_issue_refs or []))
-        context_text = "\n".join(
-            [
-                "===== Changed files (name-status) =====",
-                "M\tdocs/features/active/fix-all-script/spec.md",
-                "===== Diff shortstat =====",
-            ]
-        )
-        return PRContextResult(
-            text=context_text,
-            referenced_issues=["#1"],
-            referenced_prs=["#2"],
-            verified_closing=["#1"],
-            invalid_references=[],
-            base_ref=base_ref,
-            resolved_base="origin/main",
-            base_sha="base-sha",
-            head_ref=head_ref or "feature",
-            head_sha="head-sha",
-            merge_base="base-sha",
-            rev_range="base-sha..head-sha",
-            gh_available=True,
-        )
-
-    monkeypatch.setattr("scripts.dev_tools.pr_context.collector.GitClient", FakeGit)
-    monkeypatch.setattr("scripts.dev_tools.pr_context.collector.GhClient", FakeGh)
-    monkeypatch.setattr(
-        "scripts.dev_tools.pr_context.collector.build_pr_context",
-        fake_build_pr_context,
-    )
-
-    def fake_gather_feature_excerpts(
-        root: Path, paths: Sequence[str]
-    ) -> list[FeatureDocExcerpt]:
-        return [
-            FeatureDocExcerpt(
-                feature="fix-all-script",
-                excerpt="Excerpt",
-                issue_refs=["#7"],
-                context_files=["docs/features/active/fix-all-script/spec.md"],
-            )
-        ]
-
-    monkeypatch.setattr(
-        "scripts.dev_tools.pr_context.collector.gather_feature_excerpts",
-        fake_gather_feature_excerpts,
-    )
-
-    collect_and_write(
-        base="main",
-        head="feature",
-        out=mem_path / "summary.txt",
-        appendix_out=mem_path / "appendix.txt",
-        repo_root=mem_path,
-        append=False,
-        include_untracked=False,
-    )
-
-    assert feature_calls[0] == []
-    assert feature_calls[1] == ["#7"]
-    summary_text = next(text for path, text in captured if path.name == "summary.txt")
-    appendix_text = next(text for path, text in captured if path.name == "appendix.txt")
-    assert "Scoping docs changed" in summary_text
-    assert "fix-all-script" in appendix_text
-
-
-def test_extract_digest_bullets_truncates_and_skips_blank_lines() -> None:
-    body = "## Why\n- first\n\n- second\n- third"
-    bullets = _extract_digest_bullets(body, headings=["Why"], limit=2)
-    assert bullets == ["Why: first", "Why: second"]
-
-
-def test_parse_numstat_detailed_skips_invalid_rows() -> None:
-    adds, dels, mapping = _parse_numstat_detailed("\n1\t1\tfile.py\ninvalid-entry")
-    assert adds == 1 and dels == 1
-    assert mapping == {"file.py": (1, 1)}
-
-
-def test_write_output_creates_parent_and_appends(mem_path: Path) -> None:
-    target = mem_path / "nested" / "out.txt"
-    write_output("first", target, append=False)
-    write_output("second", target, append=True)
-    assert target.read_text(encoding="utf-8").endswith("firstsecond")
-
-
-def test_parse_args_and_main_delegate_to_collect(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    captured: dict[str, object] = {}
-
-    def fake_collect_and_write(**kwargs: object) -> None:
-        captured.update(kwargs)
-
-    monkeypatch.setattr(
-        "scripts.dev_tools.pr_context.collector.collect_and_write",
-        fake_collect_and_write,
-    )
-
-    args = parse_args(
-        [
-            "--base",
-            "origin/dev",
-            "--head",
-            "feature/123",
-            "--out",
-            "~/summary.txt",
-            "--appendix-out",
-            "~/appendix.txt",
-            "--repo-root",
-            ".",
-            "--append",
-            "--no-untracked",
-        ]
-    )
-    assert args.base == "origin/dev"
-    assert args.append is True
-
-    main(
-        [
-            "--base",
-            "origin/dev",
-            "--head",
-            "feature/123",
-            "--out",
-            "~/summary.txt",
-            "--appendix-out",
-            "~/appendix.txt",
-            "--repo-root",
-            ".",
-            "--append",
-            "--no-untracked",
-        ]
-    )
-
-    assert captured["base"] == "origin/dev"
-    assert captured["head"] == "feature/123"
-    assert captured["append"] is True
-    assert captured["include_untracked"] is False
-
-
-def test_collect_and_write_renders_non_material_scoping(
-    monkeypatch: pytest.MonkeyPatch, mem_path: Path
-) -> None:
-    outputs: list[tuple[Path, str]] = []
-
-    def fake_write_output(text: str, out_path: Path, append: bool) -> None:
-        outputs.append((out_path, text))
-
-    monkeypatch.setattr(
-        "scripts.dev_tools.pr_context.collector.write_output", fake_write_output
-    )
-
-    class StubGit:
-        def __init__(self, *args: object, **kwargs: object) -> None:
-            self._root = mem_path
-
-        def resolve_root(self) -> Path:
-            return self._root
-
-        def branch_name(self) -> str:
-            return "feature/ABC-10"
-
-        def upstream(self) -> str:
-            return "origin/feature/ABC-10"
-
-        def remote_verbose(self) -> str:
-            return "origin https://example/repo (fetch)"
-
-        def status_short(self) -> str:
-            return "## feature/ABC-10"
-
-        def untracked(self) -> str:
-            return ""
-
-        def diff_name_status(self, *, staged: bool) -> str:
-            return ""
-
-        def diff_patch(self, *, staged: bool) -> str:
-            return ""
-
-        def rev_parse(self, ref: str) -> str:
-            return f"{ref}-sha"
-
-        def merge_base(self, base: str, head: str) -> str:
-            return "base-sha"
-
-        def log(self, fmt: str, rev_range: str) -> str:
-            return ""
-
-        def diff_range(self, args: Sequence[str]) -> str:
-            if "--name-status" in args:
-                return "M\tdocs/features/active/feat/spec.md\nR100\told.py\tnew.py"
-            if "--numstat" in args:
-                return "2\t1\tdocs/features/active/feat/spec.md\n3\t0\tcore.py"
-            if "--shortstat" in args:
-                return " 1 files changed, 2 insertions(+), 1 deletions(-)"
-            if "--stat" in args:
-                return " spec.md | 3 ++"
-            return ""
-
-        def run(
-            self, args: Sequence[str], *, allow_error: bool = False
-        ) -> CommandResult:
-            return CommandResult(stdout="resolved", stderr="", code=0)
-
-    class StubGh:
-        status_message = "ok"
-
-        def __init__(self, *args: object, **kwargs: object) -> None:
-            self.available = True
-
-        def ensure_available(self) -> None:
-            return None
-
-        def classify_entity(self, number: str) -> str | None:
-            return "issue" if number == "1" else None
-
-        def issue_details(self, number: str) -> IssueDetails:
-            return IssueDetails(
-                number=f"#{number}",
-                title="Issue",
-                state="open",
-                labels=[],
-                assignees=[],
-                author="alex",
-                created_at="2024-01-01",
-                updated_at="2024-01-02",
-                body="Body",
-                comments=[],
-            )
-
-        def pr_details(self, number: str) -> PullRequestDetails:
-            return PullRequestDetails(
-                number=f"#{number}",
-                title="PR",
-                state="open",
-                author="alex",
-                base_ref="main",
-                head_ref="feature",
-                created_at="2024-01-01",
-                updated_at="2024-01-02",
-                merged_at=None,
-                labels=[],
-                assignees=[],
-                body="PR body",
-                closing_issues=["#1"],
-                files_changed=["file.py"],
-            )
-
-        def ci_status(self, head_sha: str) -> tuple[str | None, list[str]]:
-            return "success", []
-
-        def current_pr(self) -> PullRequestDetails | None:
-            return PullRequestDetails(
-                number="#10",
-                title="existing",
-                state="open",
-                author="alex",
-                base_ref="main",
-                head_ref="feature",
-                created_at="2024-01-01",
-                updated_at="2024-01-02",
-                merged_at=None,
-                labels=[],
-                assignees=[],
-                body="existing body",
-                closing_issues=["#1"],
-                files_changed=[],
-            )
-
-    def fake_build_pr_context(
-        *,
-        git: StubGit,
-        gh: StubGh,
-        base_ref: str | None,
-        head_ref: str | None,
-        include_untracked: bool,
-        feature_issue_refs: Sequence[str] | None = None,
-        current_pr: PullRequestDetails | None = None,
-        gh_available: bool | None = None,
-    ) -> PRContextResult:
-        context_text = "\n".join(
-            [
-                "===== Changed files (name-status) =====",
-                "M\tdocs/features/active/feat/spec.md",
-                "A\tcore.py",
-                "===== Diff shortstat =====",
-                " 1 files changed, 2 insertions(+), 1 deletions(-)",
-            ]
-        )
-        return PRContextResult(
-            text=context_text,
-            referenced_issues=["#1"],
-            referenced_prs=["#2"],
-            verified_closing=["#1"],
-            invalid_references=["#999"],
-            base_ref=base_ref,
-            resolved_base="origin/main",
-            base_sha="base-sha",
-            head_ref=head_ref or "feature",
-            head_sha="head-sha",
-            merge_base="base-sha",
-            rev_range="base-sha..head-sha",
-            gh_available=True,
-        )
-
-    monkeypatch.setattr("scripts.dev_tools.pr_context.collector.GitClient", StubGit)
-    monkeypatch.setattr("scripts.dev_tools.pr_context.collector.GhClient", StubGh)
-    monkeypatch.setattr(
-        "scripts.dev_tools.pr_context.collector.build_pr_context", fake_build_pr_context
-    )
-
-    def fake_gather_feature_excerpts(
-        *args: object, **kwargs: object
-    ) -> list[FeatureDocExcerpt]:
-        return []
-
-    def fake_scoping_doc_changes(
-        **kwargs: object,
-    ) -> list[tuple[str, bool, list[str], None]]:
-        return [("docs/features/active/feat/spec.md", False, ["links only"], None)]
-
-    monkeypatch.setattr(
-        "scripts.dev_tools.pr_context.collector.gather_feature_excerpts",
-        fake_gather_feature_excerpts,
-    )
-    monkeypatch.setattr(
-        "scripts.dev_tools.pr_context.collector._scoping_doc_changes",
-        fake_scoping_doc_changes,
-    )
-
-    collect_and_write(
-        base="main",
-        head="feature",
-        out=mem_path / "summary.txt",
-        appendix_out=mem_path / "appendix.txt",
-        repo_root=mem_path,
-        append=False,
-        include_untracked=False,
-    )
-
-    summary_text = next(text for path, text in outputs if path.name == "summary.txt")
-    assert "non-material" in summary_text
-    assert "#999" in summary_text
-
-
-def test_collect_and_write_handles_offline_gh(
-    monkeypatch: pytest.MonkeyPatch, mem_path: Path
-) -> None:
-    outputs: list[tuple[Path, str]] = []
-
-    def fake_write_output(text: str, out_path: Path, append: bool) -> None:
-        outputs.append((out_path, text))
-
-    monkeypatch.setattr(
-        "scripts.dev_tools.pr_context.collector.write_output", fake_write_output
-    )
-
-    class OfflineGit:
-        def __init__(self, *args: object, **kwargs: object) -> None:
-            self._root = mem_path
-
-        def resolve_root(self) -> Path:
-            return self._root
-
-        def branch_name(self) -> str:
-            return "feature/ABC-10"
-
-        def upstream(self) -> str:
-            return "origin/feature/ABC-10"
-
-        def remote_verbose(self) -> str:
-            return "origin https://example/repo (fetch)"
-
-        def status_short(self) -> str:
-            return "## feature/ABC-10"
-
-        def untracked(self) -> str:
-            return ""
-
-        def diff_name_status(self, *, staged: bool) -> str:
-            return ""
-
-        def diff_patch(self, *, staged: bool) -> str:
-            return ""
-
-        def rev_parse(self, ref: str) -> str:
-            return f"{ref}-sha"
-
-        def merge_base(self, base: str, head: str) -> str:
-            return "base-sha"
-
-        def log(self, fmt: str, rev_range: str) -> str:
-            return ""
-
-        def diff_range(self, args: Sequence[str]) -> str:
-            return ""
-
-        def run(
-            self, args: Sequence[str], *, allow_error: bool = False
-        ) -> CommandResult:
-            return CommandResult(stdout="resolved", stderr="", code=0)
-
-    class OfflineGh:
-        status_message = None
-        available = False
-
-        def __init__(self, *args: object, **kwargs: object) -> None:
-            return None
-
-        def ensure_available(self) -> None:
-            raise RuntimeError("offline")
-
-        def current_pr(self) -> PullRequestDetails | None:
-            return None
-
-        def classify_entity(self, number: str) -> str | None:
-            return None
-
-    monkeypatch.setattr("scripts.dev_tools.pr_context.collector.GitClient", OfflineGit)
-    monkeypatch.setattr("scripts.dev_tools.pr_context.collector.GhClient", OfflineGh)
-
-    collect_and_write(
-        base="main",
-        head="feature",
-        out=mem_path / "summary.txt",
-        appendix_out=mem_path / "appendix.txt",
-        repo_root=mem_path,
-        append=False,
-        include_untracked=False,
-    )
-
-    summary_text = next(text for path, text in outputs if path.name == "summary.txt")
-    assert "GitHub CLI unavailable" in summary_text or "unavailable" in summary_text
-    normalized_summary = summary_text.lower().replace("-", "")
-    assert "autoclose issues" in normalized_summary
-
-
-def test_collect_and_write_includes_intent_and_additional_context(
-    monkeypatch: pytest.MonkeyPatch, mem_path: Path
-) -> None:
-    outputs: list[tuple[Path, str]] = []
-
-    def fake_write_output(text: str, out_path: Path, append: bool) -> None:
-        outputs.append((out_path, text))
-
-    monkeypatch.setattr(
-        "scripts.dev_tools.pr_context.collector.write_output", fake_write_output
-    )
-
-    class StubGit:
-        def __init__(self, *args: object, **kwargs: object) -> None:
-            self._root = Path(__file__).resolve().parents[3]
-
-        def resolve_root(self) -> Path:
-            return self._root
-
-        def branch_name(self) -> str:
-            return "feature/ABC-10"
-
-        def upstream(self) -> str:
-            return "origin/feature/ABC-10"
-
-        def remote_verbose(self) -> str:
-            return "origin https://example/repo (fetch)"
-
-        def status_short(self) -> str:
-            return "## feature/ABC-10"
-
-        def untracked(self) -> str:
-            return ""
-
-        def diff_name_status(self, *, staged: bool) -> str:
-            return ""
-
-        def diff_patch(self, *, staged: bool) -> str:
-            return ""
-
-        def rev_parse(self, ref: str) -> str:
-            return f"{ref}-sha"
-
-        def merge_base(self, base: str, head: str) -> str:
-            return "base-sha"
-
-        def log(self, fmt: str, rev_range: str) -> str:
-            return ""
-
-        def diff_range(self, args: Sequence[str]) -> str:
-            if "--name-status" in args:
-                return "M\tdocs/features/active/2025-12-18-docs-v3-upgrade/spec.md"
-            if "--numstat" in args:
-                return "1\t0\tdocs/features/active/2025-12-18-docs-v3-upgrade/spec.md"
-            if "--shortstat" in args:
-                return " 1 files changed, 1 insertions(+), 0 deletions(-)"
-            if "--stat" in args:
-                return " spec.md | 1 +"
-            return ""
-
-        def run(
-            self, args: Sequence[str], *, allow_error: bool = False
-        ) -> CommandResult:
-            return CommandResult(stdout="resolved", stderr="", code=0)
-
-    class StubGh:
-        status_message = "ok"
-        available = True
-
-        def __init__(self, *args: object, **kwargs: object) -> None:
-            return None
-
-        def ensure_available(self) -> None:
-            return None
-
-        def current_pr(self) -> PullRequestDetails | None:
-            return None
-
-        def classify_entity(self, number: str) -> str | None:
-            return None
-
-        def ci_status(self, head_sha: str) -> tuple[str | None, list[str]]:
-            return "success", []
-
-    def fake_build_pr_context(
-        *,
-        git: StubGit,
-        gh: StubGh,
-        base_ref: str | None,
-        head_ref: str | None,
-        include_untracked: bool,
-        feature_issue_refs: Sequence[str] | None = None,
-        current_pr: PullRequestDetails | None = None,
-        gh_available: bool | None = None,
-    ) -> PRContextResult:
-        context_text = "\n".join(
-            [
-                "===== Changed files (name-status) =====",
-                "M\tdocs/features/active/2025-12-18-docs-v3-upgrade/spec.md",
-                "===== Diff shortstat =====",
-                " 1 files changed, 1 insertions(+), 0 deletions(-)",
-            ]
-        )
-        return PRContextResult(
-            text=context_text,
-            referenced_issues=[],
-            referenced_prs=[],
-            verified_closing=[],
-            invalid_references=[],
-            base_ref=base_ref,
-            resolved_base="origin/main",
-            base_sha="base-sha",
-            head_ref=head_ref or "feature",
-            head_sha="head-sha",
-            merge_base="base-sha",
-            rev_range="base-sha..head-sha",
-            gh_available=True,
-        )
-
-    monkeypatch.setattr("scripts.dev_tools.pr_context.collector.GitClient", StubGit)
-    monkeypatch.setattr("scripts.dev_tools.pr_context.collector.GhClient", StubGh)
-    monkeypatch.setattr(
-        "scripts.dev_tools.pr_context.collector.build_pr_context", fake_build_pr_context
-    )
-
-    def fake_gather_feature_excerpts(
-        root: Path, paths: Sequence[str]
-    ) -> list[FeatureDocExcerpt]:
-        return [
-            FeatureDocExcerpt(
-                feature="2025-12-18-docs-v3-upgrade",
-                excerpt=(
-                    "Feature doc: 2025-12-18-docs-v3-upgrade\n"
-                    "Plan verification notes\n"
-                    "Excerpt"
-                ),
-                issue_refs=[],
-                context_files=[
-                    "docs/features/active/2025-12-18-docs-v3-upgrade/spec.md",
-                    "docs/features/active/2025-12-18-docs-v3-upgrade/user-story.md",
-                ],
-            )
-        ]
-
-    monkeypatch.setattr(
-        "scripts.dev_tools.pr_context.collector.gather_feature_excerpts",
-        fake_gather_feature_excerpts,
-    )
-
-    repo_root = Path(__file__).resolve().parents[3]
-    collect_and_write(
-        base="main",
-        head="feature",
-        out=mem_path / "summary.txt",
-        appendix_out=mem_path / "appendix.txt",
-        repo_root=repo_root,
-        append=False,
-        include_untracked=False,
-    )
-
-    summary_text = next(text for path, text in outputs if path.name == "summary.txt")
-    appendix_text = next(text for path, text in outputs if path.name == "appendix.txt")
-
-    assert "PR Intent" in summary_text
-    assert "Author-asserted autoclose issues" in summary_text
-    assert "Additional context files" in summary_text
-    assert "Feature doc excerpts" in summary_text
-    assert "Excerpt" in summary_text
-    assert "Feature: 2025-12-18-docs-v3-upgrade" in summary_text
-    assert "Context files:" in summary_text
-    assert (
-        "docs/features/active/2025-12-18-docs-v3-upgrade/user-story.md" in summary_text
-    )
-    assert "Feature doc: 2025-12-18-docs-v3-upgrade" in appendix_text
-    assert "Plan verification notes" in appendix_text
