+++ tests/scripts/dev_tools/test_collect_pr_context_part3.py
@@ -1,0 +1,414 @@
+from __future__ import annotations
+
+from pathlib import Path
+from typing import TYPE_CHECKING
+
+import pytest
+
+if TYPE_CHECKING:
+    from collections.abc import Sequence
+
+from scripts.dev_tools.pr_context.collector import (
+    CommandResult,
+    GitClient,
+    IssueDetails,
+    PullRequestDetails,
+    collect_and_write,
+)
+from scripts.dev_tools.pr_context.models import FeatureDocExcerpt, PRContextResult
+
+
+@pytest.fixture
+def mem_path(tmp_path: Path) -> Path:
+    """Alias fixture for cosmetic tmp_path->mem_path test parameter rename."""
+    return tmp_path
+
+
+class FakeRunner:
+    def __init__(self, responses: dict[tuple[str, ...], CommandResult]) -> None:
+        self.responses = responses
+
+    def run(
+        self,
+        args: Sequence[str],
+        *,
+        cwd: Path | None = None,
+        allow_error: bool = False,
+    ) -> CommandResult:
+        key = tuple(args)
+        return self.responses.get(key, CommandResult(stdout="", stderr="", code=1))
+
+
+class FakeGit(GitClient):
+    def __init__(self) -> None:
+        super().__init__(FakeRunner({}), Path("."))
+        self.called = []
+
+    def branch_name(self) -> str:
+        return "feature/test"
+
+    def upstream(self) -> str:
+        return "origin/feature/test"
+
+    def remote_verbose(self) -> str:
+        return "origin https://example/repo (fetch)"
+
+    def status_short(self) -> str:
+        return "## feature/test...origin/feature/test"
+
+    def untracked(self) -> str:
+        return "docs/features/active/fix-all-script/spec.md"
+
+    def diff_name_status(self, *, staged: bool) -> str:
+        marker = "A" if staged else "M"
+        return f"{marker}\tdocs/features/active/fix-all-script/spec.md"
+
+    def diff_patch(self, *, staged: bool) -> str:
+        return "(diff omitted)"
+
+    def rev_parse(self, ref: str) -> str:
+        return "basehash" if ref == "main" else "headhash"
+
+    def merge_base(self, base: str, head: str) -> str:
+        return base
+
+    def log(self, fmt: str, rev_range: str) -> str:
+        if fmt.startswith("--pretty=format:"):
+            return "\n".join(
+                [
+                    "a1 2025-01-01 alice Merge pull request #53",
+                    "a2 2025-01-02 bob fix(scope): closes #44",
+                ]
+            )
+        if fmt == "--pretty=%s":
+            return "Merge pull request #53\nfix(scope): closes #44"
+        if fmt == "--format=%an <%ae>":
+            return "alice <a@example.com>\nbob <b@example.com>"
+        return ""
+
+    def diff_range(self, args: Sequence[str]) -> str:
+        if "--name-status" in args:
+            return "M\tdocs/features/active/fix-all-script/spec.md"
+        if "--numstat" in args:
+            return "4\t2\tdocs/features/active/fix-all-script/spec.md"
+        if "--shortstat" in args:
+            return " 1 files changed, 4 insertions(+), 2 deletions(-)"
+        if "--stat" in args:
+            return " spec.md | 6 +-"
+        return ""
+
+    def run(self, args: Sequence[str], *, allow_error: bool = False) -> CommandResult:
+        return CommandResult(stdout="resolved", stderr="", code=0)
+
+
+class FakeGh:
+    def __init__(self) -> None:
+        self.available = True
+
+    def ensure_available(self) -> None:
+        return None
+
+    def classify_entity(self, number: str) -> str | None:
+        if number == "44":
+            return "issue"
+        if number == "53":
+            return "pull"
+        return None
+
+
+def test_collect_and_write_renders_non_material_scoping(
+    monkeypatch: pytest.MonkeyPatch, mem_path: Path
+) -> None:
+    outputs: list[tuple[Path, str]] = []
+
+    def fake_write_output(text: str, out_path: Path, append: bool) -> None:
+        outputs.append((out_path, text))
+
+    monkeypatch.setattr(
+        "scripts.dev_tools.pr_context.collector.write_output", fake_write_output
+    )
+
+    class StubGit:
+        def __init__(self, *args: object, **kwargs: object) -> None:
+            self._root = mem_path
+
+        def resolve_root(self) -> Path:
+            return self._root
+
+        def branch_name(self) -> str:
+            return "feature/ABC-10"
+
+        def upstream(self) -> str:
+            return "origin/feature/ABC-10"
+
+        def remote_verbose(self) -> str:
+            return "origin https://example/repo (fetch)"
+
+        def status_short(self) -> str:
+            return "## feature/ABC-10"
+
+        def untracked(self) -> str:
+            return ""
+
+        def diff_name_status(self, *, staged: bool) -> str:
+            return ""
+
+        def diff_patch(self, *, staged: bool) -> str:
+            return ""
+
+        def rev_parse(self, ref: str) -> str:
+            return f"{ref}-sha"
+
+        def merge_base(self, base: str, head: str) -> str:
+            return "base-sha"
+
+        def log(self, fmt: str, rev_range: str) -> str:
+            return ""
+
+        def diff_range(self, args: Sequence[str]) -> str:
+            if "--name-status" in args:
+                return "M\tdocs/features/active/feat/spec.md\nR100\told.py\tnew.py"
+            if "--numstat" in args:
+                return "2\t1\tdocs/features/active/feat/spec.md\n3\t0\tcore.py"
+            if "--shortstat" in args:
+                return " 1 files changed, 2 insertions(+), 1 deletions(-)"
+            if "--stat" in args:
+                return " spec.md | 3 ++"
+            return ""
+
+        def run(
+            self, args: Sequence[str], *, allow_error: bool = False
+        ) -> CommandResult:
+            return CommandResult(stdout="resolved", stderr="", code=0)
+
+    class StubGh:
+        status_message = "ok"
+
+        def __init__(self, *args: object, **kwargs: object) -> None:
+            self.available = True
+
+        def ensure_available(self) -> None:
+            return None
+
+        def classify_entity(self, number: str) -> str | None:
+            return "issue" if number == "1" else None
+
+        def issue_details(self, number: str) -> IssueDetails:
+            return IssueDetails(
+                number=f"#{number}",
+                title="Issue",
+                state="open",
+                labels=[],
+                assignees=[],
+                author="alex",
+                created_at="2024-01-01",
+                updated_at="2024-01-02",
+                body="Body",
+                comments=[],
+            )
+
+        def pr_details(self, number: str) -> PullRequestDetails:
+            return PullRequestDetails(
+                number=f"#{number}",
+                title="PR",
+                state="open",
+                author="alex",
+                base_ref="main",
+                head_ref="feature",
+                created_at="2024-01-01",
+                updated_at="2024-01-02",
+                merged_at=None,
+                labels=[],
+                assignees=[],
+                body="PR body",
+                closing_issues=["#1"],
+                files_changed=["file.py"],
+            )
+
+        def ci_status(self, head_sha: str) -> tuple[str | None, list[str]]:
+            return "success", []
+
+        def current_pr(self) -> PullRequestDetails | None:
+            return PullRequestDetails(
+                number="#10",
+                title="existing",
+                state="open",
+                author="alex",
+                base_ref="main",
+                head_ref="feature",
+                created_at="2024-01-01",
+                updated_at="2024-01-02",
+                merged_at=None,
+                labels=[],
+                assignees=[],
+                body="existing body",
+                closing_issues=["#1"],
+                files_changed=[],
+            )
+
+    def fake_build_pr_context(
+        *,
+        git: StubGit,
+        gh: StubGh,
+        base_ref: str | None,
+        head_ref: str | None,
+        include_untracked: bool,
+        feature_issue_refs: Sequence[str] | None = None,
+        current_pr: PullRequestDetails | None = None,
+        gh_available: bool | None = None,
+    ) -> PRContextResult:
+        context_text = "\n".join(
+            [
+                "===== Changed files (name-status) =====",
+                "M\tdocs/features/active/feat/spec.md",
+                "A\tcore.py",
+                "===== Diff shortstat =====",
+                " 1 files changed, 2 insertions(+), 1 deletions(-)",
+            ]
+        )
+        return PRContextResult(
+            text=context_text,
+            referenced_issues=["#1"],
+            referenced_prs=["#2"],
+            verified_closing=["#1"],
+            invalid_references=["#999"],
+            base_ref=base_ref,
+            resolved_base="origin/main",
+            base_sha="base-sha",
+            head_ref=head_ref or "feature",
+            head_sha="head-sha",
+            merge_base="base-sha",
+            rev_range="base-sha..head-sha",
+            gh_available=True,
+        )
+
+    monkeypatch.setattr("scripts.dev_tools.pr_context.collector.GitClient", StubGit)
+    monkeypatch.setattr("scripts.dev_tools.pr_context.collector.GhClient", StubGh)
+    monkeypatch.setattr(
+        "scripts.dev_tools.pr_context.collector.build_pr_context", fake_build_pr_context
+    )
+
+    def fake_gather_feature_excerpts(
+        *args: object, **kwargs: object
+    ) -> list[FeatureDocExcerpt]:
+        return []
+
+    def fake_scoping_doc_changes(
+        **kwargs: object,
+    ) -> list[tuple[str, bool, list[str], None]]:
+        return [("docs/features/active/feat/spec.md", False, ["links only"], None)]
+
+    monkeypatch.setattr(
+        "scripts.dev_tools.pr_context.collector.gather_feature_excerpts",
+        fake_gather_feature_excerpts,
+    )
+    monkeypatch.setattr(
+        "scripts.dev_tools.pr_context.collector._scoping_doc_changes",
+        fake_scoping_doc_changes,
+    )
+
+    collect_and_write(
+        base="main",
+        head="feature",
+        out=mem_path / "summary.txt",
+        appendix_out=mem_path / "appendix.txt",
+        repo_root=mem_path,
+        append=False,
+        include_untracked=False,
+    )
+
+    summary_text = next(text for path, text in outputs if path.name == "summary.txt")
+    assert "non-material" in summary_text
+    assert "#999" in summary_text
+
+
+def test_collect_and_write_handles_offline_gh(
+    monkeypatch: pytest.MonkeyPatch, mem_path: Path
+) -> None:
+    outputs: list[tuple[Path, str]] = []
+
+    def fake_write_output(text: str, out_path: Path, append: bool) -> None:
+        outputs.append((out_path, text))
+
+    monkeypatch.setattr(
+        "scripts.dev_tools.pr_context.collector.write_output", fake_write_output
+    )
+
+    class OfflineGit:
+        def __init__(self, *args: object, **kwargs: object) -> None:
+            self._root = mem_path
+
+        def resolve_root(self) -> Path:
+            return self._root
+
+        def branch_name(self) -> str:
+            return "feature/ABC-10"
+
+        def upstream(self) -> str:
+            return "origin/feature/ABC-10"
+
+        def remote_verbose(self) -> str:
+            return "origin https://example/repo (fetch)"
+
+        def status_short(self) -> str:
+            return "## feature/ABC-10"
+
+        def untracked(self) -> str:
+            return ""
+
+        def diff_name_status(self, *, staged: bool) -> str:
+            return ""
+
+        def diff_patch(self, *, staged: bool) -> str:
+            return ""
+
+        def rev_parse(self, ref: str) -> str:
+            return f"{ref}-sha"
+
+        def merge_base(self, base: str, head: str) -> str:
+            return "base-sha"
+
+        def log(self, fmt: str, rev_range: str) -> str:
+            return ""
+
+        def diff_range(self, args: Sequence[str]) -> str:
+            return ""
+
+        def run(
+            self, args: Sequence[str], *, allow_error: bool = False
+        ) -> CommandResult:
+            return CommandResult(stdout="resolved", stderr="", code=0)
+
+    class OfflineGh:
+        status_message = None
+        available = False
+
+        def __init__(self, *args: object, **kwargs: object) -> None:
+            return None
+
+        def ensure_available(self) -> None:
+            raise RuntimeError("offline")
+
+        def current_pr(self) -> PullRequestDetails | None:
+            return None
+
+        def classify_entity(self, number: str) -> str | None:
+            return None
+
+    monkeypatch.setattr("scripts.dev_tools.pr_context.collector.GitClient", OfflineGit)
+    monkeypatch.setattr("scripts.dev_tools.pr_context.collector.GhClient", OfflineGh)
+
+    collect_and_write(
+        base="main",
+        head="feature",
+        out=mem_path / "summary.txt",
+        appendix_out=mem_path / "appendix.txt",
+        repo_root=mem_path,
+        append=False,
+        include_untracked=False,
+    )
+
+    summary_text = next(text for path, text in outputs if path.name == "summary.txt")
+    assert "GitHub CLI unavailable" in summary_text or "unavailable" in summary_text
+    normalized_summary = summary_text.lower().replace("-", "")
+    assert "autoclose issues" in normalized_summary
