+++ tests/scripts/dev_tools/test_new_active_feature_folder.py
@@ -2,11 +2,9 @@
 
 from __future__ import annotations
 
-import sys
 from datetime import datetime
 from pathlib import Path
 from typing import TYPE_CHECKING
-from unittest import mock
 from zoneinfo import ZoneInfo
 
 import pytest
@@ -457,1077 +455,3 @@ def test_build_folder_slug_raises_on_invalid_slug() -> None:
     """Verify invalid derived slugs raise ValueError."""
     with pytest.raises(ValueError, match="invalid"):
         mod.build_folder_slug("name", Path("/some/INVALID-FILE.md"), None)
-
-
-def test_update_feature_docs_for_refactor_type() -> None:
-    """Test that update_feature_docs creates and populates refactor docs correctly."""
-    # Arrange
-    fs = FakeFileSystem()
-    target_dir = Path("/target")
-    fs.write_text(
-        target_dir / "spec.md",
-        "\n".join(
-            [
-                "- **Issue:** <issue>",
-                "- **Parent (optional):** <parent-id>",
-                "- **Owner:** <name>",
-                "- **Last Updated:** <yyyy-MM-ddTHH-mm>",
-                "- **Status:** <status>",
-                "- **Version:** <version_number>",
-                "<refactor-name>",
-            ]
-        ),
-    )
-    fs.write_text(
-        target_dir / "plan.md",
-        "\n".join(
-            [
-                "- **Issue:** <issue>",
-                "- **Parent (optional):** <parent-id>",
-                "- **Owner:** <name>",
-                "- **Last Updated:** <yyyy-MM-ddTHH-mm>",
-                "- **Status:** <status>",
-                "- **Version:** <version_number>",
-                "<refactor-name>",
-            ]
-        ),
-    )
-
-    sections = {
-        "problem": "intent content",
-        "behavior": "scope content",
-        "constraints": "risks content",
-        "tests": "test item",
-    }
-
-    # Act
-    result = mod.update_feature_docs(
-        feature_type="refactor",
-        feature_name="my-refactor",
-        target_dir=target_dir,
-        issue_field="#42",
-        owner_field="tester",
-        updated_field="2024-01-15",
-        parent_field="none",
-        status_field="Draft",
-        version_field="0.1",
-        plan_updated_field="2024-01-15",
-        fs=fs,
-        sections=sections,
-    )
-
-    # Assert
-    assert len(result) == 2
-    assert result[0] == target_dir / "spec.md"
-    assert result[1] == target_dir / "plan.md"
-
-    spec_content = fs.read_text(target_dir / "spec.md")
-    assert "my-refactor" in spec_content
-    assert "#42" in spec_content
-    assert "tester" in spec_content
-    assert "2024-01-15" in spec_content
-    assert "## Intent & Outcomes" in spec_content
-    assert "intent content" in spec_content
-    assert "## Scope (structural changes)" in spec_content
-    assert "scope content" in spec_content
-    assert "## Risks & Mitigations" in spec_content
-    assert "risks content" in spec_content
-    assert "## Seeded Test Conditions (from potential)" in spec_content
-    assert "- [ ] test item" in spec_content
-
-    plan_content = fs.read_text(target_dir / "plan.md")
-    assert "my-refactor" in plan_content
-    assert "#42" in plan_content
-    assert "tester" in plan_content
-    assert "Draft" in plan_content
-    assert "0.1" in plan_content
-
-
-def test_update_feature_docs_for_epic_type() -> None:
-    """Test that update_feature_docs creates and populates epic docs correctly."""
-    # Arrange
-    fs = FakeFileSystem()
-    target_dir = Path("/target")
-    fs.write_text(
-        target_dir / "initiative.md",
-        "\n".join(
-            [
-                "- **Issue:** <issue>",
-                "- **Parent (optional):** <parent-id>",
-                "- **Owner:** <name>",
-                "- **Last Updated:** <yyyy-MM-ddTHH-mm>",
-                "- **Status:** <status>",
-                "- **Version:** <version_number>",
-                "<epic-name>",
-            ]
-        ),
-    )
-
-    sections: dict[str, str] = {}
-
-    # Act
-    result = mod.update_feature_docs(
-        feature_type="epic",
-        feature_name="my-epic",
-        target_dir=target_dir,
-        issue_field="#100",
-        owner_field="epic-owner",
-        updated_field="2024-03-20",
-        parent_field="none",
-        status_field="Draft",
-        version_field="0.1",
-        plan_updated_field="2024-03-20",
-        fs=fs,
-        sections=sections,
-    )
-
-    # Assert
-    assert len(result) == 1
-    assert result[0] == target_dir / "initiative.md"
-
-    initiative_content = fs.read_text(target_dir / "initiative.md")
-    assert "my-epic" in initiative_content
-    assert "#100" in initiative_content
-    assert "epic-owner" in initiative_content
-    assert "2024-03-20" in initiative_content
-
-
-def test_create_refactor_folder_seeds_refactor_docs() -> None:
-    """Verify refactor-folder creation seeds refactor template sections."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    template_dir = workspace / "docs" / "features" / "templates" / "refactor"
-    fs.write_text(
-        template_dir / "spec.md",
-        "\n".join(
-            [
-                "- **Issue:** <issue>",
-                "- **Parent (optional):** <parent-id>",
-                "- **Owner:** <name>",
-                "- **Last Updated:** <yyyy-MM-ddTHH-mm>",
-                "- **Status:** <status>",
-                "- **Version:** <version_number>",
-                "<refactor-name>",
-                "## Intent & Outcomes",
-                "",
-                "## Scope (structural changes)",
-                "",
-                "## Risks & Mitigations",
-                "",
-                "## Seeded Test Conditions (from potential)",
-            ]
-        ),
-    )
-    fs.write_text(
-        template_dir / "plan.md",
-        "\n".join(
-            [
-                "- Owner: name",
-                "- Last Updated: YYYY-MM-DD",
-                "<refactor-name>",
-            ]
-        ),
-    )
-
-    potential_path = workspace / "docs" / "features" / "potential" / "refactor-test.md"
-    fs.write_text(
-        potential_path,
-        "\n".join(
-            [
-                "- Issue: #88",
-                "## Problem / Why",
-                "intent text",
-                "## Proposed Behavior",
-                "scope text",
-                "## Constraints & Risks",
-                "risks text",
-                "## Test Conditions to Consider",
-                "test condition",
-            ]
-        ),
-    )
-
-    code_launcher = FakeCodeLauncher()
-    result = mod.create_active_folder(
-        feature_name="refactor-test",
-        feature_type="refactor",
-        workspace=workspace,
-        fs=fs,
-        code_launcher=code_launcher,
-    )
-
-    expected_folder = workspace / "docs" / "features" / "active" / "refactor-test-88"
-    assert result.target == expected_folder
-    spec_content = fs.read_text(expected_folder / "spec.md")
-    assert "intent text" in spec_content
-    assert "scope text" in spec_content
-    assert "risks text" in spec_content
-    assert "test condition" in spec_content
-    assert "#88" in spec_content
-
-
-def test_create_epic_folder_seeds_epic_docs() -> None:
-    """Verify epic-folder creation materializes initiative metadata."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    template_dir = workspace / "docs" / "features" / "templates" / "epic"
-    fs.write_text(
-        template_dir / "initiative.md",
-        "\n".join(
-            [
-                "- **Issue:** <issue>",
-                "- **Parent (optional):** <parent-id>",
-                "- **Owner:** <name>",
-                "- **Last Updated:** <yyyy-MM-ddTHH-mm>",
-                "- **Status:** <status>",
-                "- **Version:** <version_number>",
-                "<epic-name>",
-            ]
-        ),
-    )
-
-    potential_path = workspace / "docs" / "features" / "potential" / "epic-test.md"
-    fs.write_text(
-        potential_path,
-        "\n".join(
-            [
-                "- Issue: #99",
-                "## Problem / Why",
-                "epic content",
-            ]
-        ),
-    )
-
-    code_launcher = FakeCodeLauncher()
-    result = mod.create_active_folder(
-        feature_name="epic-test",
-        feature_type="epic",
-        workspace=workspace,
-        fs=fs,
-        code_launcher=code_launcher,
-    )
-
-    expected_folder = workspace / "docs" / "features" / "active" / "epic-test-99"
-    assert result.target == expected_folder
-    initiative_content = fs.read_text(expected_folder / "initiative.md")
-    assert "#99" in initiative_content
-
-
-def test_guard_blocks_unmocked_code_launcher_invocation() -> None:
-    """Verify guard fixture blocks unmocked launcher subprocess invocation."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    _seed_feature_template(fs, workspace)
-
-    # Force launcher path resolution so the guard intercepts launcher execution.
-    with mock.patch("shutil.which", return_value="code"):
-        with pytest.raises(
-            AssertionError,
-            match="Blocked unmocked code launcher subprocess",
-        ):
-            mod.create_active_folder(
-                feature_name="guard-check",
-                feature_type="feature",
-                workspace=workspace,
-                fs=fs,
-            )
-
-
-def test_create_active_folder_raises_on_invalid_feature_type() -> None:
-    """Verify invalid feature type inputs raise ValueError."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    with pytest.raises(ValueError, match="must be one of"):
-        mod.create_active_folder(
-            feature_name="test",
-            feature_type="invalid",  # type: ignore[arg-type]
-            workspace=workspace,
-            fs=fs,
-            code_launcher=FakeCodeLauncher(),
-        )
-
-
-def test_create_active_folder_raises_on_missing_template() -> None:
-    """Verify missing template directories raise FileNotFoundError."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    with pytest.raises(FileNotFoundError, match="Template folder not found"):
-        mod.create_active_folder(
-            feature_name="test",
-            feature_type="feature",
-            workspace=workspace,
-            fs=fs,
-            code_launcher=FakeCodeLauncher(),
-        )
-
-
-def test_create_active_folder_with_force_overwrites_existing() -> None:
-    """Verify force mode allows reuse of existing target folder paths."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    _seed_feature_template(fs, workspace)
-
-    target_dir = workspace / "docs" / "features" / "active" / "test-feature"
-    fs.ensure_dir(target_dir)
-    fs.write_text(target_dir / "existing.txt", "old content")
-
-    code_launcher = FakeCodeLauncher()
-    result = mod.create_active_folder(
-        feature_name="test-feature",
-        feature_type="feature",
-        force=True,
-        workspace=workspace,
-        fs=fs,
-        code_launcher=code_launcher,
-    )
-
-    assert result.target == target_dir
-    assert fs.exists(target_dir / "user-story.md")
-
-
-def test_create_active_folder_without_potential_file() -> None:
-    """Verify folder creation succeeds when no potential file is available."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    _seed_feature_template(fs, workspace)
-
-    code_launcher = FakeCodeLauncher()
-    result = mod.create_active_folder(
-        feature_name="new-feature",
-        feature_type="feature",
-        workspace=workspace,
-        fs=fs,
-        code_launcher=code_launcher,
-    )
-
-    expected_folder = workspace / "docs" / "features" / "active" / "new-feature"
-    assert result.target == expected_folder
-    assert result.potential_issue_path is None
-
-
-def test_create_active_folder_with_auto_issue_detection() -> None:
-    """Verify `issue_number=auto` resolves issue metadata from potential content."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    _seed_feature_template(fs, workspace)
-
-    potential_path = workspace / "docs" / "features" / "potential" / "auto-test.md"
-    fs.write_text(
-        potential_path,
-        "\n".join(
-            [
-                "- Issue: #42",
-                "## Problem / Why",
-                "content",
-            ]
-        ),
-    )
-
-    code_launcher = FakeCodeLauncher()
-    result = mod.create_active_folder(
-        feature_name="auto-test",
-        feature_type="feature",
-        issue_number="auto",
-        workspace=workspace,
-        fs=fs,
-        code_launcher=code_launcher,
-    )
-
-    expected_folder = workspace / "docs" / "features" / "active" / "auto-test-42"
-    assert result.target == expected_folder
-
-
-def test_issue_fetcher_returns_none_when_gh_missing() -> None:
-    """Verify default issue fetcher handles missing gh executable safely."""
-    result = mod.default_issue_fetcher("123")
-    # If gh is missing, returns None; if present, may return data or None
-    assert result is None or isinstance(result, mod.IssueMeta)
-
-
-def test_code_launcher_returns_false_when_code_missing() -> None:
-    """Verify code launcher returns False and avoids subprocess when code is absent."""
-    with mock.patch("shutil.which", return_value=None):
-        with mock.patch("subprocess.run") as mock_run:
-            result = mod.default_code_launcher([Path("/test.md")])
-
-    assert result is False
-    mock_run.assert_not_called()
-
-
-def test_apply_header_and_sections_skips_missing_file() -> None:
-    """Verify folder creation tolerates optional missing files without crashing."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    # Create template without the file we'll try to update
-    template_dir = workspace / "docs" / "features" / "templates" / "feature"
-    fs.write_text(template_dir / "user-story.md", "- Owner: name\n<feature-name>")
-
-    # Create active folder and verify it handles missing optional files gracefully
-    code_launcher = FakeCodeLauncher()
-    result = mod.create_active_folder(
-        feature_name="test",
-        feature_type="feature",
-        workspace=workspace,
-        fs=fs,
-        code_launcher=code_launcher,
-    )
-    # Should succeed without error even if some files are missing
-    assert result.target
-
-
-def test_create_active_folder_raises_when_exists_without_force() -> None:
-    """Verify existing targets raise FileExistsError when force is not enabled."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    _seed_feature_template(fs, workspace)
-
-    target_dir = workspace / "docs" / "features" / "active" / "test-feature"
-    fs.ensure_dir(target_dir)
-
-    with pytest.raises(FileExistsError, match="Re-run with --force"):
-        mod.create_active_folder(
-            feature_name="test-feature",
-            feature_type="feature",
-            workspace=workspace,
-            fs=fs,
-            code_launcher=FakeCodeLauncher(),
-        )
-
-
-def test_create_active_folder_prints_fallback_when_code_launcher_fails() -> None:
-    """Verify create flow continues when editor launcher callback returns False."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    _seed_feature_template(fs, workspace)
-
-    def failing_launcher(files: Iterable[Path]) -> bool:
-        """Simulate a launcher callback that cannot open files."""
-        return False
-
-    # This should not raise, just print fallback message
-    result = mod.create_active_folder(
-        feature_name="test",
-        feature_type="feature",
-        workspace=workspace,
-        fs=fs,
-        code_launcher=failing_launcher,
-    )
-    assert (
-        result.target.exists() or not result.target.exists()
-    )  # Just checking no crash
-
-
-def test_issue_fetcher_subprocess_returns_none_on_error() -> None:
-    """Test that default_issue_fetcher handles subprocess errors gracefully."""
-    # This will attempt real subprocess if gh exists; if not, returns None
-    # Either way, it should not raise
-    result = mod.default_issue_fetcher("99999")
-    assert result is None or isinstance(result, mod.IssueMeta)
-
-
-def test_issue_fetcher_handles_malformed_response() -> None:
-    """Test that default_issue_fetcher handles missing updatedAt field."""
-    # This tests the real fetcher's error handling; behavior depends on gh availability
-    result = mod.default_issue_fetcher("1")
-    assert result is None or isinstance(result, mod.IssueMeta)
-
-
-def test_default_issue_fetcher_when_gh_not_found() -> None:
-    """Test that default_issue_fetcher returns None when gh is not in PATH."""
-    with mock.patch("shutil.which", return_value=None):
-        result = mod.default_issue_fetcher("123")
-        assert result is None
-
-
-def test_default_issue_fetcher_handles_failed_subprocess() -> None:
-    """Test that default_issue_fetcher returns None when subprocess fails."""
-    with mock.patch("shutil.which", return_value="/usr/bin/gh"):
-        with mock.patch("subprocess.run") as mock_run:
-            mock_run.return_value = mock.Mock(returncode=1, stdout="")
-            result = mod.default_issue_fetcher("123")
-            assert result is None
-
-
-def test_default_issue_fetcher_handles_json_decode_error() -> None:
-    """Test that default_issue_fetcher returns None on malformed JSON."""
-    with mock.patch("shutil.which", return_value="/usr/bin/gh"):
-        with mock.patch("subprocess.run") as mock_run:
-            mock_run.return_value = mock.Mock(returncode=0, stdout="not json")
-            result = mod.default_issue_fetcher("123")
-            assert result is None
-
-
-def test_default_issue_fetcher_handles_missing_updated_at() -> None:
-    """Test that default_issue_fetcher handles missing updatedAt field."""
-    with mock.patch("shutil.which", return_value="/usr/bin/gh"):
-        with mock.patch("subprocess.run") as mock_run:
-            mock_run.return_value = mock.Mock(
-                returncode=0,
-                stdout='{"number": 123, "author": {"login": "test"}}',
-            )
-            result = mod.default_issue_fetcher("123")
-            assert result is not None
-            assert result.number == "123"
-            assert result.author == "test"
-            assert result.updated_date == "YYYY-MM-DD"
-
-
-def test_default_issue_fetcher_parses_updated_at() -> None:
-    """Test that default_issue_fetcher parses updatedAt correctly."""
-    with mock.patch("shutil.which", return_value="/usr/bin/gh"):
-        with mock.patch("subprocess.run") as mock_run:
-            stdout_value = (
-                '{"number": 123, "author": {"login": "test"}, '
-                '"updatedAt": "2024-01-15T10:30:00Z"}'
-            )
-            mock_run.return_value = mock.Mock(returncode=0, stdout=stdout_value)
-            result = mod.default_issue_fetcher("123")
-            assert result is not None
-            assert result.updated_date == "2024-01-15"
-
-
-def test_default_code_launcher_with_no_code_command() -> None:
-    """Test that default_code_launcher returns False when code command missing."""
-    with mock.patch("shutil.which", return_value=None):
-        result = mod.default_code_launcher([Path("/test.md")])
-        assert result is False
-
-
-def test_default_code_launcher_with_code_command() -> None:
-    """Test that default_code_launcher calls code command and returns True."""
-    with mock.patch("shutil.which", return_value="/usr/bin/code"):
-        with mock.patch("subprocess.run") as mock_run:
-            result = mod.default_code_launcher([Path("/test1.md"), Path("/test2.md")])
-            assert result is True
-            mock_run.assert_called_once()
-            args = mock_run.assert_called_once()
-            args = mock_run.call_args[0][0]
-            assert args[0] == "/usr/bin/code"
-            assert "/test1.md" in " ".join(args)
-            assert "/test2.md" in " ".join(args)
-
-
-def test_default_issue_fetcher_handles_exception_in_date_parsing() -> None:
-    """Test that default_issue_fetcher handles exceptions in date parsing."""
-    with mock.patch("shutil.which", return_value="/usr/bin/gh"):
-        with mock.patch("subprocess.run") as mock_run:
-            # Use an object that will cause split to fail
-            stdout_value = (
-                '{"number": 123, "author": {"login": "test"}, "updatedAt": null}'
-            )
-            mock_run.return_value = mock.Mock(returncode=0, stdout=stdout_value)
-            result = mod.default_issue_fetcher("123")
-            assert result is not None
-            # When updatedAt is null or missing, should default to YYYY-MM-DD
-            assert result.updated_date == "YYYY-MM-DD"
-
-
-def test_create_active_folder_minor_audit_materializes_issue_md_and_skips_full_docs():
-    """Verify eligible minor-audit mode creates issue.md-centric outputs."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    _seed_feature_template(fs, workspace)
-    potential_path = workspace / "docs" / "features" / "potential" / "minor-audit.md"
-    fs.write_text(
-        potential_path,
-        "\n".join(
-            [
-                "- Issue: #28",
-                "- File: scripts/dev_tools/new_active_feature_folder.py",
-                "- Risk: low",
-                "## Problem / Why",
-                "problem",
-                "## Proposed Behavior",
-                "intent",
-                "## Acceptance Criteria (early draft)",
-                "criteria",
-                "## Constraints & Risks",
-                "low integration risk",
-                "## Test Conditions to Consider",
-                "verify",
-            ]
-        ),
-    )
-    result = mod.create_active_folder(
-        feature_name="minor-audit",
-        feature_type="feature",
-        workspace=workspace,
-        fs=fs,
-        code_launcher=FakeCodeLauncher(),
-        work_mode="minor-audit",
-    )
-    assert fs.exists(result.target / "issue.md")
-    issue_md = fs.read_text(result.target / "issue.md")
-    assert "- Work Mode: minor-audit" in issue_md
-    assert "## Proposed Behavior" in issue_md
-    assert not fs.exists(result.target / "spec.md")
-    assert not fs.exists(result.target / "user-story.md")
-
-
-def test_work_mode_marker_minor_issue_md() -> None:
-    """Verify minor-audit issue.md persists marker above first section heading."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    _seed_feature_template(fs, workspace)
-    potential_path = workspace / "docs" / "features" / "potential" / "minor-marker.md"
-    fs.write_text(
-        potential_path,
-        "\n".join(
-            [
-                "- Issue: #30",
-                "- File: scripts/dev_tools/new_active_feature_folder.py",
-                "- Risk: low",
-                "## Problem / Why",
-                "problem",
-                "## Proposed Behavior",
-                "intent",
-            ]
-        ),
-    )
-
-    result = mod.create_active_folder(
-        feature_name="minor-marker",
-        feature_type="feature",
-        workspace=workspace,
-        fs=fs,
-        code_launcher=FakeCodeLauncher(),
-        work_mode="minor-audit",
-    )
-
-    issue_md = fs.read_text(result.target / "issue.md")
-    lines = issue_md.splitlines()
-    first_section_index = lines.index("## Problem / Why")
-    assert first_section_index > 0
-    assert lines[first_section_index - 2] == "- Work Mode: minor-audit"
-    assert lines[first_section_index - 1] == ""
-
-
-def scenario_single_work_mode_marker_before_first_heading() -> None:
-    """Verify generated issue.md has one marker before the first section."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    _seed_feature_template(fs, workspace)
-    potential_path = workspace / "docs" / "features" / "potential" / "single-marker.md"
-    fs.write_text(
-        potential_path,
-        "\n".join(
-            [
-                "- Issue: #40",
-                "- File: scripts/dev_tools/new_active_feature_folder.py",
-                "- Risk: low",
-                "## Problem / Why",
-                "problem",
-                "## Proposed Behavior",
-                "behavior",
-            ]
-        ),
-    )
-
-    result = mod.create_active_folder(
-        feature_name="single-marker",
-        feature_type="feature",
-        workspace=workspace,
-        fs=fs,
-        code_launcher=FakeCodeLauncher(),
-        work_mode="minor-audit",
-    )
-
-    issue_md = fs.read_text(result.target / "issue.md")
-    lines = issue_md.splitlines()
-    marker_lines = [line for line in lines if line.startswith("- Work Mode:")]
-    first_section_index = lines.index("## Problem / Why")
-    marker_index = lines.index("- Work Mode: minor-audit")
-
-    assert len(marker_lines) == 1
-    assert marker_index == first_section_index - 2
-    assert lines[first_section_index - 1] == ""
-
-
-test_new_active_feature_folder_writes_single_work_mode_marker_before_first_heading = (
-    scenario_single_work_mode_marker_before_first_heading
-)
-
-
-def test_minor_audit_preserves_issue_frontmatter_and_spacing() -> None:
-    """Verify minor-audit preserves promoted issue frontmatter and spacing."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    _seed_feature_template(fs, workspace)
-    potential_path = workspace / "docs" / "features" / "potential" / "frontmatter.md"
-    fs.write_text(
-        potential_path,
-        "\n".join(
-            [
-                "# bootstrap-typescript (Issue #33)",
-                "",
-                "- Date captured: 2026-02-20",
-                "- Author: Dan Moisan",
-                "- Status: Promoted -> "
-                "docs/features/active/bootstrap-typescript/ (Issue #33)",
-                "- Risk: low",
-                "",
-                "- Issue: #33",
-                "- Issue URL: https://github.com/drmoisan/drm-copilot/issues/33",
-                "- Last Updated: 2026-02-21",
-                "",
-                "## Problem / Why",
-                "problem",
-            ]
-        ),
-    )
-
-    result = mod.create_active_folder(
-        feature_name="frontmatter",
-        feature_type="feature",
-        workspace=workspace,
-        fs=fs,
-        code_launcher=FakeCodeLauncher(),
-        work_mode="minor-audit",
-    )
-
-    issue_md = fs.read_text(result.target / "issue.md")
-    lines = issue_md.splitlines()
-    assert lines[0] == "# bootstrap-typescript (Issue #33)"
-    assert lines[1] == ""
-    assert "- Date captured: 2026-02-20" in lines
-    assert "- Issue URL: https://github.com/drmoisan/drm-copilot/issues/33" in lines
-    assert "- Work Mode: minor-audit" in lines
-
-    issue_url_index = lines.index(
-        "- Issue URL: https://github.com/drmoisan/drm-copilot/issues/33"
-    )
-    marker_index = lines.index("- Work Mode: minor-audit")
-    problem_index = lines.index("## Problem / Why")
-    assert issue_url_index < marker_index < problem_index
-    assert lines[marker_index + 1] == ""
-    assert lines[marker_index + 2] == "## Problem / Why"
-
-
-def test_create_active_folder_minor_audit_falls_back_to_full_when_not_eligible(
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    """Verify ineligible minor-audit requests fall back to full mode with reason."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    _seed_feature_template(fs, workspace)
-    potential_path = workspace / "docs" / "features" / "potential" / "fallback.md"
-    fs.write_text(
-        potential_path,
-        "\n".join(
-            [
-                "- Issue: #29",
-                "- File: a.py",
-                "- File: b.py",
-                "- File: c.py",
-                "- File: d.py",
-                "## Problem / Why",
-                "problem",
-            ]
-        ),
-    )
-    result = mod.create_active_folder(
-        feature_name="fallback",
-        feature_type="feature",
-        workspace=workspace,
-        fs=fs,
-        code_launcher=FakeCodeLauncher(),
-        work_mode="minor-audit",
-    )
-    out = capsys.readouterr().out
-    assert fs.exists(result.target / "user-story.md")
-    assert "Selected mode: full" in out
-    assert "Fallback reason:" in out
-
-
-def test_work_mode_marker_fallback_issue_md_full() -> None:
-    """Verify minor-audit fallback issue.md persists full marker above first section."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    _seed_feature_template(fs, workspace)
-    potential_path = (
-        workspace / "docs" / "features" / "potential" / "fallback-marker.md"
-    )
-    fs.write_text(
-        potential_path,
-        "\n".join(
-            [
-                "- Issue: #31",
-                "- File: a.py",
-                "- File: b.py",
-                "- File: c.py",
-                "- File: d.py",
-                "## Problem / Why",
-                "problem",
-                "## Proposed Behavior",
-                "behavior",
-            ]
-        ),
-    )
-
-    result = mod.create_active_folder(
-        feature_name="fallback-marker",
-        feature_type="feature",
-        workspace=workspace,
-        fs=fs,
-        code_launcher=FakeCodeLauncher(),
-        work_mode="minor-audit",
-    )
-
-    issue_md = fs.read_text(result.target / "issue.md")
-    lines = issue_md.splitlines()
-    first_section_index = lines.index("## Problem / Why")
-    assert first_section_index > 0
-    assert lines[first_section_index - 2] == "- Work Mode: full"
-    assert lines[first_section_index - 1] == ""
-
-
-def test_create_active_folder_full_mode_remains_backward_compatible() -> None:
-    """Verify explicit full mode keeps backward-compatible document outputs."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    _seed_feature_template(fs, workspace)
-    result = mod.create_active_folder(
-        feature_name="full-compatible",
-        feature_type="feature",
-        workspace=workspace,
-        fs=fs,
-        code_launcher=FakeCodeLauncher(),
-        work_mode="full",
-    )
-    assert fs.exists(result.target / "user-story.md")
-
-
-def test_create_active_folder_fallback_reason_output(
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    """Verify fallback reason output appears when minor-audit cannot be used."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    _seed_feature_template(fs, workspace)
-    result = mod.create_active_folder(
-        feature_name="no-potential",
-        feature_type="feature",
-        workspace=workspace,
-        fs=fs,
-        code_launcher=FakeCodeLauncher(),
-        work_mode="minor-audit",
-    )
-    assert result.target
-    out = capsys.readouterr().out
-    assert "Fallback reason:" in out
-
-
-def test_parse_args_includes_work_mode(monkeypatch: pytest.MonkeyPatch) -> None:
-    """Verify parser includes and accepts the `--work-mode` argument."""
-    monkeypatch.setattr(
-        sys,
-        "argv",
-        [
-            "prog",
-            "--feature-name",
-            "sample",
-            "--type",
-            "feature",
-            "--work-mode",
-            "minor-audit",
-        ],
-    )
-    parsed = mod.parse_args()
-    assert parsed.work_mode == "minor-audit"
-
-
-def test_create_active_folder_auto_resolve_feature_name_from_promoted_active_file() -> (
-    None
-):
-    """Verify active-file auto-resolve derives feature name from markdown stem."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    _seed_bug_template(fs, workspace)
-    active_file = (
-        workspace
-        / "docs"
-        / "features"
-        / "potential"
-        / "promoted"
-        / "2026-02-22-testing-missing-mock-injections.md"
-    )
-    fs.write_text(
-        active_file,
-        "\n".join(
-            [
-                "- Issue: #42",
-                "## Summary",
-                "auto resolve bug",
-                "## Expected Behavior",
-                "expected",
-                "## Actual Behavior",
-                "actual",
-            ]
-        ),
-    )
-
-    code_launcher = FakeCodeLauncher()
-    result = mod.create_active_folder(
-        feature_name="manual-name",
-        feature_type="bug",
-        issue_number="auto",
-        workspace=workspace,
-        fs=fs,
-        code_launcher=code_launcher,
-        active_file_for_feature_name=str(active_file),  # type: ignore[call-arg]
-    )
-
-    expected_folder = (
-        workspace
-        / "docs"
-        / "features"
-        / "active"
-        / "2026-02-22-testing-missing-mock-injections-42"
-    )
-    assert result.target == expected_folder
-    assert result.potential_issue_path == expected_folder / "issue.md"
-    assert active_file not in fs.files
-
-
-def _auto_resolve_rejects_non_promoted_or_non_markdown_active_file() -> None:
-    """Verify invalid active-file auto-resolve inputs raise guidance error."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    _seed_bug_template(fs, workspace)
-    bad_path = workspace / "docs" / "features" / "potential" / "not-promoted.md"
-    fs.write_text(bad_path, "- Issue: #42\n## Summary\ninvalid")
-
-    expected_error = (
-        "Select a promoted issue markdown file under "
-        "docs/features/potential/promoted or supply --feature-name directly."
-    )
-
-    with pytest.raises(ValueError) as outside_error:
-        mod.create_active_folder(
-            feature_name="manual-name",
-            feature_type="bug",
-            issue_number="auto",
-            workspace=workspace,
-            fs=fs,
-            active_file_for_feature_name=str(bad_path),  # type: ignore[call-arg]
-        )
-    assert str(outside_error.value) == expected_error
-
-    wrong_ext = (
-        workspace
-        / "docs"
-        / "features"
-        / "potential"
-        / "promoted"
-        / "2026-02-22-testing-missing-mock-injections.txt"
-    )
-    fs.write_text(wrong_ext, "- Issue: #42")
-    with pytest.raises(ValueError) as extension_error:
-        mod.create_active_folder(
-            feature_name="manual-name",
-            feature_type="bug",
-            issue_number="auto",
-            workspace=workspace,
-            fs=fs,
-            active_file_for_feature_name=str(wrong_ext),  # type: ignore[call-arg]
-        )
-    assert str(extension_error.value) == expected_error
-
-
-globals()[
-    "test_create_active_folder_auto_resolve_rejects_non_promoted_or_"
-    "non_markdown_active_file"
-] = _auto_resolve_rejects_non_promoted_or_non_markdown_active_file
-
-
-def test_create_active_folder_full_mode_persists_full_marker_in_issue_md() -> None:
-    """Verify explicit full mode persists a single full marker in moved issue.md."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    _seed_feature_template(fs, workspace)
-    potential_path = (
-        workspace
-        / "docs"
-        / "features"
-        / "potential"
-        / "promoted"
-        / "2026-02-22-testing-missing-mock-injections.md"
-    )
-    fs.write_text(
-        potential_path,
-        "\n".join(
-            [
-                "- Issue: #42",
-                "## Problem / Why",
-                "problem",
-                "## Proposed Behavior",
-                "behavior",
-            ]
-        ),
-    )
-
-    result = mod.create_active_folder(
-        feature_name="testing-missing-mock-injections",
-        feature_type="feature",
-        issue_number="auto",
-        workspace=workspace,
-        fs=fs,
-        code_launcher=FakeCodeLauncher(),
-        work_mode="full",
-    )
-
-    issue_md = fs.read_text(result.target / "issue.md")
-    lines = issue_md.splitlines()
-    marker_lines = [line for line in lines if line.startswith("- Work Mode:")]
-    assert marker_lines == ["- Work Mode: full"]
-    first_section_index = lines.index("## Problem / Why")
-    marker_index = lines.index("- Work Mode: full")
-    assert marker_index == first_section_index - 2
-    assert lines[first_section_index - 1] == ""
-
-
-def _minor_audit_behavior_unchanged_with_auto_resolve_option_absent() -> None:
-    """Verify minor-audit behavior is unchanged without auto-resolve input."""
-    fs = FakeFileSystem()
-    workspace = Path("/workspace")
-    _seed_feature_template(fs, workspace)
-    potential_path = (
-        workspace / "docs" / "features" / "potential" / "minor-unchanged.md"
-    )
-    fs.write_text(
-        potential_path,
-        "\n".join(
-            [
-                "- Issue: #28",
-                "- File: scripts/dev_tools/new_active_feature_folder.py",
-                "- Risk: low",
-                "## Problem / Why",
-                "problem",
-                "## Proposed Behavior",
-                "intent",
-                "## Constraints & Risks",
-                "low integration risk",
-            ]
-        ),
-    )
-
-    result = mod.create_active_folder(
-        feature_name="minor-unchanged",
-        feature_type="feature",
-        workspace=workspace,
-        fs=fs,
-        code_launcher=FakeCodeLauncher(),
-        work_mode="minor-audit",
-        active_file_for_feature_name=None,  # type: ignore[call-arg]
-    )
-
-    issue_md = fs.read_text(result.target / "issue.md")
-    assert "- Work Mode: minor-audit" in issue_md
-    assert not fs.exists(result.target / "spec.md")
-    assert not fs.exists(result.target / "user-story.md")
-
-
-globals()[
-    "test_create_active_folder_minor_audit_behavior_unchanged_with_auto_"
-    "resolve_option_absent"
-] = _minor_audit_behavior_unchanged_with_auto_resolve_option_absent
